refactoring the first samples
