+++ 101-linq-samples/src/Restrictions.cs
@@ -1,5414 +1,14 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Xml.Linq;
 
 namespace Try101LinqSamples
 {
-    public class Product
-    {
-        public int ProductID { get; set; }
-        public string ProductName { get; set; }
-        public string Category { get; set; }
-        public decimal UnitPrice { get; set; }
-        public int UnitsInStock { get; set; }
-    }
-
-    public class Customer
-    {
-        public string CustomerID { get; set; }
-        public string CompanyName { get; set; }
-        public string Address { get; set; }
-        public string City { get; set; }
-        public string Region { get; set; }
-        public string PostalCode { get; set; }
-        public string Country { get; set; }
-        public string Phone { get; set; }
-        public string Fax { get; set; }
-        public Order[] Orders { get; set; }
-    }
-
-    public class Order
-    {
-        public int OrderID { get; set; }
-        public DateTime OrderDate { get; set; }
-        public decimal Total { get; set; }
-    }
-
-
     public class Restrictions
     {
-        #region XML_DATA // You don't really want to see this.
-        private const string customers_xml =
-@"<?xml version=""1.0""?>
-<customers>
-  <customer>
-    <id>ALFKI</id>
-    <name>Alfreds Futterkiste</name>
-    <address>Obere Str. 57</address>
-    <city>Berlin</city>
-    <postalcode>12209</postalcode>
-    <country>Germany</country>
-    <phone>030-0074321</phone>
-    <fax>030-0076545</fax>
-    <orders>
-      <order>
-        <id>10643</id>
-        <orderdate>1997-08-25T00:00:00</orderdate>
-        <total>814.50</total>
-      </order>
-      <order>
-        <id>10692</id>
-        <orderdate>1997-10-03T00:00:00</orderdate>
-        <total>878.00</total>
-      </order>
-      <order>
-        <id>10702</id>
-        <orderdate>1997-10-13T00:00:00</orderdate>
-        <total>330.00</total>
-      </order>
-      <order>
-        <id>10835</id>
-        <orderdate>1998-01-15T00:00:00</orderdate>
-        <total>845.80</total>
-      </order>
-      <order>
-        <id>10952</id>
-        <orderdate>1998-03-16T00:00:00</orderdate>
-        <total>471.20</total>
-      </order>
-      <order>
-        <id>11011</id>
-        <orderdate>1998-04-09T00:00:00</orderdate>
-        <total>933.50</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>ANATR</id>
-    <name>Ana Trujillo Emparedados y helados</name>
-    <address>Avda.de la Constitución 2222</address>
-    <city>México D.F.</city>
-    <postalcode>05021</postalcode>
-    <country>Mexico</country>
-    <phone>(5) 555-4729</phone>
-    <fax>(5) 555-3745</fax>
-    <orders>
-      <order>
-        <id>10308</id>
-        <orderdate>1996-09-18T00:00:00</orderdate>
-        <total>88.80</total>
-      </order>
-      <order>
-        <id>10625</id>
-        <orderdate>1997-08-08T00:00:00</orderdate>
-        <total>479.75</total>
-      </order>
-      <order>
-        <id>10759</id>
-        <orderdate>1997-11-28T00:00:00</orderdate>
-        <total>320.00</total>
-      </order>
-      <order>
-        <id>10926</id>
-        <orderdate>1998-03-04T00:00:00</orderdate>
-        <total>514.40</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>ANTON</id>
-    <name>Antonio Moreno Taquería</name>
-    <address>Mataderos  2312</address>
-    <city>México D.F.</city>
-    <postalcode>05023</postalcode>
-    <country>Mexico</country>
-    <phone>(5) 555-3932</phone>
-    <orders>
-      <order>
-        <id>10365</id>
-        <orderdate>1996-11-27T00:00:00</orderdate>
-        <total>403.20</total>
-      </order>
-      <order>
-        <id>10507</id>
-        <orderdate>1997-04-15T00:00:00</orderdate>
-        <total>749.06</total>
-      </order>
-      <order>
-        <id>10535</id>
-        <orderdate>1997-05-13T00:00:00</orderdate>
-        <total>1940.85</total>
-      </order>
-      <order>
-        <id>10573</id>
-        <orderdate>1997-06-19T00:00:00</orderdate>
-        <total>2082.00</total>
-      </order>
-      <order>
-        <id>10677</id>
-        <orderdate>1997-09-22T00:00:00</orderdate>
-        <total>813.36</total>
-      </order>
-      <order>
-        <id>10682</id>
-        <orderdate>1997-09-25T00:00:00</orderdate>
-        <total>375.50</total>
-      </order>
-      <order>
-        <id>10856</id>
-        <orderdate>1998-01-28T00:00:00</orderdate>
-        <total>660.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>AROUT</id>
-    <name>Around the Horn</name>
-    <address>120 Hanover Sq.</address>
-    <city>London</city>
-    <postalcode>WA1 1DP</postalcode>
-    <country>UK</country>
-    <phone>(171) 555-7788</phone>
-    <fax>(171) 555-6750</fax>
-    <orders>
-      <order>
-        <id>10355</id>
-        <orderdate>1996-11-15T00:00:00</orderdate>
-        <total>480.00</total>
-      </order>
-      <order>
-        <id>10383</id>
-        <orderdate>1996-12-16T00:00:00</orderdate>
-        <total>899.00</total>
-      </order>
-      <order>
-        <id>10453</id>
-        <orderdate>1997-02-21T00:00:00</orderdate>
-        <total>407.70</total>
-      </order>
-      <order>
-        <id>10558</id>
-        <orderdate>1997-06-04T00:00:00</orderdate>
-        <total>2142.90</total>
-      </order>
-      <order>
-        <id>10707</id>
-        <orderdate>1997-10-16T00:00:00</orderdate>
-        <total>1641.00</total>
-      </order>
-      <order>
-        <id>10741</id>
-        <orderdate>1997-11-14T00:00:00</orderdate>
-        <total>228.00</total>
-      </order>
-      <order>
-        <id>10743</id>
-        <orderdate>1997-11-17T00:00:00</orderdate>
-        <total>319.20</total>
-      </order>
-      <order>
-        <id>10768</id>
-        <orderdate>1997-12-08T00:00:00</orderdate>
-        <total>1477.00</total>
-      </order>
-      <order>
-        <id>10793</id>
-        <orderdate>1997-12-24T00:00:00</orderdate>
-        <total>191.10</total>
-      </order>
-      <order>
-        <id>10864</id>
-        <orderdate>1998-02-02T00:00:00</orderdate>
-        <total>282.00</total>
-      </order>
-      <order>
-        <id>10920</id>
-        <orderdate>1998-03-03T00:00:00</orderdate>
-        <total>390.00</total>
-      </order>
-      <order>
-        <id>10953</id>
-        <orderdate>1998-03-16T00:00:00</orderdate>
-        <total>4441.25</total>
-      </order>
-      <order>
-        <id>11016</id>
-        <orderdate>1998-04-10T00:00:00</orderdate>
-        <total>491.50</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>BERGS</id>
-    <name>Berglunds snabbköp</name>
-    <address>Berguvsvägen  8</address>
-    <city>Luleå</city>
-    <postalcode>S-958 22</postalcode>
-    <country>Sweden</country>
-    <phone>0921-12 34 65</phone>
-    <fax>0921-12 34 67</fax>
-    <orders>
-      <order>
-        <id>10278</id>
-        <orderdate>1996-08-12T00:00:00</orderdate>
-        <total>1488.80</total>
-      </order>
-      <order>
-        <id>10280</id>
-        <orderdate>1996-08-14T00:00:00</orderdate>
-        <total>613.20</total>
-      </order>
-      <order>
-        <id>10384</id>
-        <orderdate>1996-12-16T00:00:00</orderdate>
-        <total>2222.40</total>
-      </order>
-      <order>
-        <id>10444</id>
-        <orderdate>1997-02-12T00:00:00</orderdate>
-        <total>1031.70</total>
-      </order>
-      <order>
-        <id>10445</id>
-        <orderdate>1997-02-13T00:00:00</orderdate>
-        <total>174.90</total>
-      </order>
-      <order>
-        <id>10524</id>
-        <orderdate>1997-05-01T00:00:00</orderdate>
-        <total>3192.65</total>
-      </order>
-      <order>
-        <id>10572</id>
-        <orderdate>1997-06-18T00:00:00</orderdate>
-        <total>1501.08</total>
-      </order>
-      <order>
-        <id>10626</id>
-        <orderdate>1997-08-11T00:00:00</orderdate>
-        <total>1503.60</total>
-      </order>
-      <order>
-        <id>10654</id>
-        <orderdate>1997-09-02T00:00:00</orderdate>
-        <total>601.83</total>
-      </order>
-      <order>
-        <id>10672</id>
-        <orderdate>1997-09-17T00:00:00</orderdate>
-        <total>3815.25</total>
-      </order>
-      <order>
-        <id>10689</id>
-        <orderdate>1997-10-01T00:00:00</orderdate>
-        <total>472.50</total>
-      </order>
-      <order>
-        <id>10733</id>
-        <orderdate>1997-11-07T00:00:00</orderdate>
-        <total>1459.00</total>
-      </order>
-      <order>
-        <id>10778</id>
-        <orderdate>1997-12-16T00:00:00</orderdate>
-        <total>96.50</total>
-      </order>
-      <order>
-        <id>10837</id>
-        <orderdate>1998-01-16T00:00:00</orderdate>
-        <total>1064.50</total>
-      </order>
-      <order>
-        <id>10857</id>
-        <orderdate>1998-01-28T00:00:00</orderdate>
-        <total>2048.21</total>
-      </order>
-      <order>
-        <id>10866</id>
-        <orderdate>1998-02-03T00:00:00</orderdate>
-        <total>1096.20</total>
-      </order>
-      <order>
-        <id>10875</id>
-        <orderdate>1998-02-06T00:00:00</orderdate>
-        <total>709.55</total>
-      </order>
-      <order>
-        <id>10924</id>
-        <orderdate>1998-03-04T00:00:00</orderdate>
-        <total>1835.70</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>BLAUS</id>
-    <name>Blauer See Delikatessen</name>
-    <address>Forsterstr. 57</address>
-    <city>Mannheim</city>
-    <postalcode>68306</postalcode>
-    <country>Germany</country>
-    <phone>0621-08460</phone>
-    <fax>0621-08924</fax>
-    <orders>
-      <order>
-        <id>10501</id>
-        <orderdate>1997-04-09T00:00:00</orderdate>
-        <total>149.00</total>
-      </order>
-      <order>
-        <id>10509</id>
-        <orderdate>1997-04-17T00:00:00</orderdate>
-        <total>136.80</total>
-      </order>
-      <order>
-        <id>10582</id>
-        <orderdate>1997-06-27T00:00:00</orderdate>
-        <total>330.00</total>
-      </order>
-      <order>
-        <id>10614</id>
-        <orderdate>1997-07-29T00:00:00</orderdate>
-        <total>464.00</total>
-      </order>
-      <order>
-        <id>10853</id>
-        <orderdate>1998-01-27T00:00:00</orderdate>
-        <total>625.00</total>
-      </order>
-      <order>
-        <id>10956</id>
-        <orderdate>1998-03-17T00:00:00</orderdate>
-        <total>677.00</total>
-      </order>
-      <order>
-        <id>11058</id>
-        <orderdate>1998-04-29T00:00:00</orderdate>
-        <total>858.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>BLONP</id>
-    <name>Blondel père et fils</name>
-    <address>24, place Kléber</address>
-    <city>Strasbourg</city>
-    <postalcode>67000</postalcode>
-    <country>France</country>
-    <phone>88.60.15.31</phone>
-    <fax>88.60.15.32</fax>
-    <orders>
-      <order>
-        <id>10265</id>
-        <orderdate>1996-07-25T00:00:00</orderdate>
-        <total>1176.00</total>
-      </order>
-      <order>
-        <id>10297</id>
-        <orderdate>1996-09-04T00:00:00</orderdate>
-        <total>1420.00</total>
-      </order>
-      <order>
-        <id>10360</id>
-        <orderdate>1996-11-22T00:00:00</orderdate>
-        <total>7390.20</total>
-      </order>
-      <order>
-        <id>10436</id>
-        <orderdate>1997-02-05T00:00:00</orderdate>
-        <total>1994.52</total>
-      </order>
-      <order>
-        <id>10449</id>
-        <orderdate>1997-02-18T00:00:00</orderdate>
-        <total>1838.20</total>
-      </order>
-      <order>
-        <id>10559</id>
-        <orderdate>1997-06-05T00:00:00</orderdate>
-        <total>520.41</total>
-      </order>
-      <order>
-        <id>10566</id>
-        <orderdate>1997-06-12T00:00:00</orderdate>
-        <total>1761.00</total>
-      </order>
-      <order>
-        <id>10584</id>
-        <orderdate>1997-06-30T00:00:00</orderdate>
-        <total>593.75</total>
-      </order>
-      <order>
-        <id>10628</id>
-        <orderdate>1997-08-12T00:00:00</orderdate>
-        <total>450.00</total>
-      </order>
-      <order>
-        <id>10679</id>
-        <orderdate>1997-09-23T00:00:00</orderdate>
-        <total>660.00</total>
-      </order>
-      <order>
-        <id>10826</id>
-        <orderdate>1998-01-12T00:00:00</orderdate>
-        <total>730.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>BOLID</id>
-    <name>Bólido Comidas preparadas</name>
-    <address>C/ Araquil, 67</address>
-    <city>Madrid</city>
-    <postalcode>28023</postalcode>
-    <country>Spain</country>
-    <phone>(91) 555 22 82</phone>
-    <fax>(91) 555 91 99</fax>
-    <orders>
-      <order>
-        <id>10326</id>
-        <orderdate>1996-10-10T00:00:00</orderdate>
-        <total>982.00</total>
-      </order>
-      <order>
-        <id>10801</id>
-        <orderdate>1997-12-29T00:00:00</orderdate>
-        <total>3026.85</total>
-      </order>
-      <order>
-        <id>10970</id>
-        <orderdate>1998-03-24T00:00:00</orderdate>
-        <total>224.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>BONAP</id>
-    <name>Bon app'</name>
-    <address>12, rue des Bouchers</address>
-    <city>Marseille</city>
-    <postalcode>13008</postalcode>
-    <country>France</country>
-    <phone>91.24.45.40</phone>
-    <fax>91.24.45.41</fax>
-    <orders>
-      <order>
-        <id>10331</id>
-        <orderdate>1996-10-16T00:00:00</orderdate>
-        <total>88.50</total>
-      </order>
-      <order>
-        <id>10340</id>
-        <orderdate>1996-10-29T00:00:00</orderdate>
-        <total>2436.18</total>
-      </order>
-      <order>
-        <id>10362</id>
-        <orderdate>1996-11-25T00:00:00</orderdate>
-        <total>1549.60</total>
-      </order>
-      <order>
-        <id>10470</id>
-        <orderdate>1997-03-11T00:00:00</orderdate>
-        <total>1820.80</total>
-      </order>
-      <order>
-        <id>10511</id>
-        <orderdate>1997-04-18T00:00:00</orderdate>
-        <total>2550.00</total>
-      </order>
-      <order>
-        <id>10525</id>
-        <orderdate>1997-05-02T00:00:00</orderdate>
-        <total>818.40</total>
-      </order>
-      <order>
-        <id>10663</id>
-        <orderdate>1997-09-10T00:00:00</orderdate>
-        <total>1930.40</total>
-      </order>
-      <order>
-        <id>10715</id>
-        <orderdate>1997-10-23T00:00:00</orderdate>
-        <total>1296.00</total>
-      </order>
-      <order>
-        <id>10730</id>
-        <orderdate>1997-11-05T00:00:00</orderdate>
-        <total>484.26</total>
-      </order>
-      <order>
-        <id>10732</id>
-        <orderdate>1997-11-06T00:00:00</orderdate>
-        <total>360.00</total>
-      </order>
-      <order>
-        <id>10755</id>
-        <orderdate>1997-11-26T00:00:00</orderdate>
-        <total>1948.50</total>
-      </order>
-      <order>
-        <id>10827</id>
-        <orderdate>1998-01-12T00:00:00</orderdate>
-        <total>843.00</total>
-      </order>
-      <order>
-        <id>10871</id>
-        <orderdate>1998-02-05T00:00:00</orderdate>
-        <total>1979.23</total>
-      </order>
-      <order>
-        <id>10876</id>
-        <orderdate>1998-02-09T00:00:00</orderdate>
-        <total>917.00</total>
-      </order>
-      <order>
-        <id>10932</id>
-        <orderdate>1998-03-06T00:00:00</orderdate>
-        <total>1788.63</total>
-      </order>
-      <order>
-        <id>10940</id>
-        <orderdate>1998-03-11T00:00:00</orderdate>
-        <total>360.00</total>
-      </order>
-      <order>
-        <id>11076</id>
-        <orderdate>1998-05-06T00:00:00</orderdate>
-        <total>792.75</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>BOTTM</id>
-    <name>Bottom-Dollar Markets</name>
-    <address>23 Tsawassen Blvd.</address>
-    <city>Tsawassen</city>
-    <region>BC</region>
-    <postalcode>T2F 8M4</postalcode>
-    <country>Canada</country>
-    <phone>(604) 555-4729</phone>
-    <fax>(604) 555-3745</fax>
-    <orders>
-      <order>
-        <id>10389</id>
-        <orderdate>1996-12-20T00:00:00</orderdate>
-        <total>1832.80</total>
-      </order>
-      <order>
-        <id>10410</id>
-        <orderdate>1997-01-10T00:00:00</orderdate>
-        <total>802.00</total>
-      </order>
-      <order>
-        <id>10411</id>
-        <orderdate>1997-01-10T00:00:00</orderdate>
-        <total>966.80</total>
-      </order>
-      <order>
-        <id>10431</id>
-        <orderdate>1997-01-30T00:00:00</orderdate>
-        <total>1892.25</total>
-      </order>
-      <order>
-        <id>10492</id>
-        <orderdate>1997-04-01T00:00:00</orderdate>
-        <total>851.20</total>
-      </order>
-      <order>
-        <id>10742</id>
-        <orderdate>1997-11-14T00:00:00</orderdate>
-        <total>3118.00</total>
-      </order>
-      <order>
-        <id>10918</id>
-        <orderdate>1998-03-02T00:00:00</orderdate>
-        <total>1447.50</total>
-      </order>
-      <order>
-        <id>10944</id>
-        <orderdate>1998-03-12T00:00:00</orderdate>
-        <total>1025.32</total>
-      </order>
-      <order>
-        <id>10949</id>
-        <orderdate>1998-03-13T00:00:00</orderdate>
-        <total>4422.00</total>
-      </order>
-      <order>
-        <id>10975</id>
-        <orderdate>1998-03-25T00:00:00</orderdate>
-        <total>717.50</total>
-      </order>
-      <order>
-        <id>10982</id>
-        <orderdate>1998-03-27T00:00:00</orderdate>
-        <total>1014.00</total>
-      </order>
-      <order>
-        <id>11027</id>
-        <orderdate>1998-04-16T00:00:00</orderdate>
-        <total>877.72</total>
-      </order>
-      <order>
-        <id>11045</id>
-        <orderdate>1998-04-23T00:00:00</orderdate>
-        <total>1309.50</total>
-      </order>
-      <order>
-        <id>11048</id>
-        <orderdate>1998-04-24T00:00:00</orderdate>
-        <total>525.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>BSBEV</id>
-    <name>B's Beverages</name>
-    <address>Fauntleroy Circus</address>
-    <city>London</city>
-    <postalcode>EC2 5NT</postalcode>
-    <country>UK</country>
-    <phone>(171) 555-1212</phone>
-    <orders>
-      <order>
-        <id>10289</id>
-        <orderdate>1996-08-26T00:00:00</orderdate>
-        <total>479.40</total>
-      </order>
-      <order>
-        <id>10471</id>
-        <orderdate>1997-03-11T00:00:00</orderdate>
-        <total>1328.00</total>
-      </order>
-      <order>
-        <id>10484</id>
-        <orderdate>1997-03-24T00:00:00</orderdate>
-        <total>386.20</total>
-      </order>
-      <order>
-        <id>10538</id>
-        <orderdate>1997-05-15T00:00:00</orderdate>
-        <total>139.80</total>
-      </order>
-      <order>
-        <id>10539</id>
-        <orderdate>1997-05-16T00:00:00</orderdate>
-        <total>355.50</total>
-      </order>
-      <order>
-        <id>10578</id>
-        <orderdate>1997-06-24T00:00:00</orderdate>
-        <total>477.00</total>
-      </order>
-      <order>
-        <id>10599</id>
-        <orderdate>1997-07-15T00:00:00</orderdate>
-        <total>493.00</total>
-      </order>
-      <order>
-        <id>10943</id>
-        <orderdate>1998-03-11T00:00:00</orderdate>
-        <total>711.00</total>
-      </order>
-      <order>
-        <id>10947</id>
-        <orderdate>1998-03-13T00:00:00</orderdate>
-        <total>220.00</total>
-      </order>
-      <order>
-        <id>11023</id>
-        <orderdate>1998-04-14T00:00:00</orderdate>
-        <total>1500.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>CACTU</id>
-    <name>Cactus Comidas para llevar</name>
-    <address>Cerrito 333</address>
-    <city>Buenos Aires</city>
-    <postalcode>1010</postalcode>
-    <country>Argentina</country>
-    <phone>(1) 135-5555</phone>
-    <fax>(1) 135-4892</fax>
-    <orders>
-      <order>
-        <id>10521</id>
-        <orderdate>1997-04-29T00:00:00</orderdate>
-        <total>225.50</total>
-      </order>
-      <order>
-        <id>10782</id>
-        <orderdate>1997-12-17T00:00:00</orderdate>
-        <total>12.50</total>
-      </order>
-      <order>
-        <id>10819</id>
-        <orderdate>1998-01-07T00:00:00</orderdate>
-        <total>477.00</total>
-      </order>
-      <order>
-        <id>10881</id>
-        <orderdate>1998-02-11T00:00:00</orderdate>
-        <total>150.00</total>
-      </order>
-      <order>
-        <id>10937</id>
-        <orderdate>1998-03-10T00:00:00</orderdate>
-        <total>644.80</total>
-      </order>
-      <order>
-        <id>11054</id>
-        <orderdate>1998-04-28T00:00:00</orderdate>
-        <total>305.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>CENTC</id>
-    <name>Centro comercial Moctezuma</name>
-    <address>Sierras de Granada 9993</address>
-    <city>México D.F.</city>
-    <postalcode>05022</postalcode>
-    <country>Mexico</country>
-    <phone>(5) 555-3392</phone>
-    <fax>(5) 555-7293</fax>
-    <orders>
-      <order>
-        <id>10259</id>
-        <orderdate>1996-07-18T00:00:00</orderdate>
-        <total>100.80</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>CHOPS</id>
-    <name>Chop-suey Chinese</name>
-    <address>Hauptstr. 29</address>
-    <city>Bern</city>
-    <postalcode>3012</postalcode>
-    <country>Switzerland</country>
-    <phone>0452-076545</phone>
-    <orders>
-      <order>
-        <id>10254</id>
-        <orderdate>1996-07-11T00:00:00</orderdate>
-        <total>556.62</total>
-      </order>
-      <order>
-        <id>10370</id>
-        <orderdate>1996-12-03T00:00:00</orderdate>
-        <total>1117.60</total>
-      </order>
-      <order>
-        <id>10519</id>
-        <orderdate>1997-04-28T00:00:00</orderdate>
-        <total>2314.20</total>
-      </order>
-      <order>
-        <id>10731</id>
-        <orderdate>1997-11-06T00:00:00</orderdate>
-        <total>1890.50</total>
-      </order>
-      <order>
-        <id>10746</id>
-        <orderdate>1997-11-19T00:00:00</orderdate>
-        <total>2311.70</total>
-      </order>
-      <order>
-        <id>10966</id>
-        <orderdate>1998-03-20T00:00:00</orderdate>
-        <total>1098.46</total>
-      </order>
-      <order>
-        <id>11029</id>
-        <orderdate>1998-04-16T00:00:00</orderdate>
-        <total>1286.80</total>
-      </order>
-      <order>
-        <id>11041</id>
-        <orderdate>1998-04-22T00:00:00</orderdate>
-        <total>1773.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>COMMI</id>
-    <name>Comércio Mineiro</name>
-    <address>Av.dos Lusíadas, 23</address>
-    <city>São Paulo</city>
-    <region>SP</region>
-    <postalcode>05432-043</postalcode>
-    <country>Brazil</country>
-    <phone>(11) 555-7647</phone>
-    <orders>
-      <order>
-        <id>10290</id>
-        <orderdate>1996-08-27T00:00:00</orderdate>
-        <total>2169.00</total>
-      </order>
-      <order>
-        <id>10466</id>
-        <orderdate>1997-03-06T00:00:00</orderdate>
-        <total>216.00</total>
-      </order>
-      <order>
-        <id>10494</id>
-        <orderdate>1997-04-02T00:00:00</orderdate>
-        <total>912.00</total>
-      </order>
-      <order>
-        <id>10969</id>
-        <orderdate>1998-03-23T00:00:00</orderdate>
-        <total>108.00</total>
-      </order>
-      <order>
-        <id>11042</id>
-        <orderdate>1998-04-22T00:00:00</orderdate>
-        <total>405.75</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>CONSH</id>
-    <name>Consolidated Holdings</name>
-    <address>Berkeley Gardens, 12 Brewery</address>
-    <city>London</city>
-    <postalcode>WX1 6LT</postalcode>
-    <country>UK</country>
-    <phone>(171) 555-2282</phone>
-    <fax>(171) 555-9199</fax>
-    <orders>
-      <order>
-        <id>10435</id>
-        <orderdate>1997-02-04T00:00:00</orderdate>
-        <total>631.60</total>
-      </order>
-      <order>
-        <id>10462</id>
-        <orderdate>1997-03-03T00:00:00</orderdate>
-        <total>156.00</total>
-      </order>
-      <order>
-        <id>10848</id>
-        <orderdate>1998-01-23T00:00:00</orderdate>
-        <total>931.50</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>DRACD</id>
-    <name>Drachenblut Delikatessen</name>
-    <address>Walserweg 21</address>
-    <city>Aachen</city>
-    <postalcode>52066</postalcode>
-    <country>Germany</country>
-    <phone>0241-039123</phone>
-    <fax>0241-059428</fax>
-    <orders>
-      <order>
-        <id>10363</id>
-        <orderdate>1996-11-26T00:00:00</orderdate>
-        <total>447.20</total>
-      </order>
-      <order>
-        <id>10391</id>
-        <orderdate>1996-12-23T00:00:00</orderdate>
-        <total>86.40</total>
-      </order>
-      <order>
-        <id>10797</id>
-        <orderdate>1997-12-25T00:00:00</orderdate>
-        <total>420.00</total>
-      </order>
-      <order>
-        <id>10825</id>
-        <orderdate>1998-01-09T00:00:00</orderdate>
-        <total>1030.76</total>
-      </order>
-      <order>
-        <id>11036</id>
-        <orderdate>1998-04-20T00:00:00</orderdate>
-        <total>1692.00</total>
-      </order>
-      <order>
-        <id>11067</id>
-        <orderdate>1998-05-04T00:00:00</orderdate>
-        <total>86.85</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>DUMON</id>
-    <name>Du monde entier</name>
-    <address>67, rue des Cinquante Otages</address>
-    <city>Nantes</city>
-    <postalcode>44000</postalcode>
-    <country>France</country>
-    <phone>40.67.88.88</phone>
-    <fax>40.67.89.89</fax>
-    <orders>
-      <order>
-        <id>10311</id>
-        <orderdate>1996-09-20T00:00:00</orderdate>
-        <total>268.80</total>
-      </order>
-      <order>
-        <id>10609</id>
-        <orderdate>1997-07-24T00:00:00</orderdate>
-        <total>424.00</total>
-      </order>
-      <order>
-        <id>10683</id>
-        <orderdate>1997-09-26T00:00:00</orderdate>
-        <total>63.00</total>
-      </order>
-      <order>
-        <id>10890</id>
-        <orderdate>1998-02-16T00:00:00</orderdate>
-        <total>860.10</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>EASTC</id>
-    <name>Eastern Connection</name>
-    <address>35 King George</address>
-    <city>London</city>
-    <postalcode>WX3 6FW</postalcode>
-    <country>UK</country>
-    <phone>(171) 555-0297</phone>
-    <fax>(171) 555-3373</fax>
-    <orders>
-      <order>
-        <id>10364</id>
-        <orderdate>1996-11-26T00:00:00</orderdate>
-        <total>950.00</total>
-      </order>
-      <order>
-        <id>10400</id>
-        <orderdate>1997-01-01T00:00:00</orderdate>
-        <total>3063.00</total>
-      </order>
-      <order>
-        <id>10532</id>
-        <orderdate>1997-05-09T00:00:00</orderdate>
-        <total>796.35</total>
-      </order>
-      <order>
-        <id>10726</id>
-        <orderdate>1997-11-03T00:00:00</orderdate>
-        <total>655.00</total>
-      </order>
-      <order>
-        <id>10987</id>
-        <orderdate>1998-03-31T00:00:00</orderdate>
-        <total>2772.00</total>
-      </order>
-      <order>
-        <id>11024</id>
-        <orderdate>1998-04-15T00:00:00</orderdate>
-        <total>1966.81</total>
-      </order>
-      <order>
-        <id>11047</id>
-        <orderdate>1998-04-24T00:00:00</orderdate>
-        <total>817.88</total>
-      </order>
-      <order>
-        <id>11056</id>
-        <orderdate>1998-04-28T00:00:00</orderdate>
-        <total>3740.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>ERNSH</id>
-    <name>Ernst Handel</name>
-    <address>Kirchgasse 6</address>
-    <city>Graz</city>
-    <postalcode>8010</postalcode>
-    <country>Austria</country>
-    <phone>7675-3425</phone>
-    <fax>7675-3426</fax>
-    <orders>
-      <order>
-        <id>10258</id>
-        <orderdate>1996-07-17T00:00:00</orderdate>
-        <total>1614.88</total>
-      </order>
-      <order>
-        <id>10263</id>
-        <orderdate>1996-07-23T00:00:00</orderdate>
-        <total>1873.80</total>
-      </order>
-      <order>
-        <id>10351</id>
-        <orderdate>1996-11-11T00:00:00</orderdate>
-        <total>5398.72</total>
-      </order>
-      <order>
-        <id>10368</id>
-        <orderdate>1996-11-29T00:00:00</orderdate>
-        <total>1689.78</total>
-      </order>
-      <order>
-        <id>10382</id>
-        <orderdate>1996-12-13T00:00:00</orderdate>
-        <total>2900.00</total>
-      </order>
-      <order>
-        <id>10390</id>
-        <orderdate>1996-12-23T00:00:00</orderdate>
-        <total>2090.88</total>
-      </order>
-      <order>
-        <id>10402</id>
-        <orderdate>1997-01-02T00:00:00</orderdate>
-        <total>2713.50</total>
-      </order>
-      <order>
-        <id>10403</id>
-        <orderdate>1997-01-03T00:00:00</orderdate>
-        <total>855.02</total>
-      </order>
-      <order>
-        <id>10430</id>
-        <orderdate>1997-01-30T00:00:00</orderdate>
-        <total>4899.20</total>
-      </order>
-      <order>
-        <id>10442</id>
-        <orderdate>1997-02-11T00:00:00</orderdate>
-        <total>1792.00</total>
-      </order>
-      <order>
-        <id>10514</id>
-        <orderdate>1997-04-22T00:00:00</orderdate>
-        <total>8623.45</total>
-      </order>
-      <order>
-        <id>10571</id>
-        <orderdate>1997-06-17T00:00:00</orderdate>
-        <total>550.59</total>
-      </order>
-      <order>
-        <id>10595</id>
-        <orderdate>1997-07-10T00:00:00</orderdate>
-        <total>4725.00</total>
-      </order>
-      <order>
-        <id>10633</id>
-        <orderdate>1997-08-15T00:00:00</orderdate>
-        <total>5510.59</total>
-      </order>
-      <order>
-        <id>10667</id>
-        <orderdate>1997-09-12T00:00:00</orderdate>
-        <total>1536.80</total>
-      </order>
-      <order>
-        <id>10698</id>
-        <orderdate>1997-10-09T00:00:00</orderdate>
-        <total>3436.44</total>
-      </order>
-      <order>
-        <id>10764</id>
-        <orderdate>1997-12-03T00:00:00</orderdate>
-        <total>2286.00</total>
-      </order>
-      <order>
-        <id>10771</id>
-        <orderdate>1997-12-10T00:00:00</orderdate>
-        <total>344.00</total>
-      </order>
-      <order>
-        <id>10773</id>
-        <orderdate>1997-12-11T00:00:00</orderdate>
-        <total>2030.40</total>
-      </order>
-      <order>
-        <id>10776</id>
-        <orderdate>1997-12-15T00:00:00</orderdate>
-        <total>6635.28</total>
-      </order>
-      <order>
-        <id>10795</id>
-        <orderdate>1997-12-24T00:00:00</orderdate>
-        <total>2158.00</total>
-      </order>
-      <order>
-        <id>10836</id>
-        <orderdate>1998-01-16T00:00:00</orderdate>
-        <total>4705.50</total>
-      </order>
-      <order>
-        <id>10854</id>
-        <orderdate>1998-01-27T00:00:00</orderdate>
-        <total>2966.50</total>
-      </order>
-      <order>
-        <id>10895</id>
-        <orderdate>1998-02-18T00:00:00</orderdate>
-        <total>6379.40</total>
-      </order>
-      <order>
-        <id>10968</id>
-        <orderdate>1998-03-23T00:00:00</orderdate>
-        <total>1408.00</total>
-      </order>
-      <order>
-        <id>10979</id>
-        <orderdate>1998-03-26T00:00:00</orderdate>
-        <total>4813.50</total>
-      </order>
-      <order>
-        <id>10990</id>
-        <orderdate>1998-04-01T00:00:00</orderdate>
-        <total>4288.85</total>
-      </order>
-      <order>
-        <id>11008</id>
-        <orderdate>1998-04-08T00:00:00</orderdate>
-        <total>4680.90</total>
-      </order>
-      <order>
-        <id>11017</id>
-        <orderdate>1998-04-13T00:00:00</orderdate>
-        <total>6750.00</total>
-      </order>
-      <order>
-        <id>11072</id>
-        <orderdate>1998-05-05T00:00:00</orderdate>
-        <total>5218.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>FAMIA</id>
-    <name>Familia Arquibaldo</name>
-    <address>Rua Orós, 92</address>
-    <city>São Paulo</city>
-    <region>SP</region>
-    <postalcode>05442-030</postalcode>
-    <country>Brazil</country>
-    <phone>(11) 555-9857</phone>
-    <orders>
-      <order>
-        <id>10347</id>
-        <orderdate>1996-11-06T00:00:00</orderdate>
-        <total>814.42</total>
-      </order>
-      <order>
-        <id>10386</id>
-        <orderdate>1996-12-18T00:00:00</orderdate>
-        <total>166.00</total>
-      </order>
-      <order>
-        <id>10414</id>
-        <orderdate>1997-01-14T00:00:00</orderdate>
-        <total>224.83</total>
-      </order>
-      <order>
-        <id>10512</id>
-        <orderdate>1997-04-21T00:00:00</orderdate>
-        <total>525.30</total>
-      </order>
-      <order>
-        <id>10581</id>
-        <orderdate>1997-06-26T00:00:00</orderdate>
-        <total>310.00</total>
-      </order>
-      <order>
-        <id>10650</id>
-        <orderdate>1997-08-29T00:00:00</orderdate>
-        <total>1779.20</total>
-      </order>
-      <order>
-        <id>10725</id>
-        <orderdate>1997-10-31T00:00:00</orderdate>
-        <total>287.80</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>FISSA</id>
-    <name>FISSA Fabrica Inter.Salchichas S.A.</name>
-    <address>C/ Moralzarzal, 86</address>
-    <city>Madrid</city>
-    <postalcode>28034</postalcode>
-    <country>Spain</country>
-    <phone>(91) 555 94 44</phone>
-    <fax>(91) 555 55 93</fax>
-    <orders />
-  </customer>
-  <customer>
-    <id>FOLIG</id>
-    <name>Folies gourmandes</name>
-    <address>184, chaussée de Tournai</address>
-    <city>Lille</city>
-    <postalcode>59000</postalcode>
-    <country>France</country>
-    <phone>20.16.10.16</phone>
-    <fax>20.16.10.17</fax>
-    <orders>
-      <order>
-        <id>10408</id>
-        <orderdate>1997-01-08T00:00:00</orderdate>
-        <total>1622.40</total>
-      </order>
-      <order>
-        <id>10480</id>
-        <orderdate>1997-03-20T00:00:00</orderdate>
-        <total>756.00</total>
-      </order>
-      <order>
-        <id>10634</id>
-        <orderdate>1997-08-15T00:00:00</orderdate>
-        <total>4985.50</total>
-      </order>
-      <order>
-        <id>10763</id>
-        <orderdate>1997-12-03T00:00:00</orderdate>
-        <total>616.00</total>
-      </order>
-      <order>
-        <id>10789</id>
-        <orderdate>1997-12-22T00:00:00</orderdate>
-        <total>3687.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>FOLKO</id>
-    <name>Folk och fä HB</name>
-    <address>Åkergatan 24</address>
-    <city>Bräcke</city>
-    <postalcode>S-844 67</postalcode>
-    <country>Sweden</country>
-    <phone>0695-34 67 21</phone>
-    <orders>
-      <order>
-        <id>10264</id>
-        <orderdate>1996-07-24T00:00:00</orderdate>
-        <total>695.62</total>
-      </order>
-      <order>
-        <id>10327</id>
-        <orderdate>1996-10-11T00:00:00</orderdate>
-        <total>1810.00</total>
-      </order>
-      <order>
-        <id>10378</id>
-        <orderdate>1996-12-10T00:00:00</orderdate>
-        <total>103.20</total>
-      </order>
-      <order>
-        <id>10434</id>
-        <orderdate>1997-02-03T00:00:00</orderdate>
-        <total>321.12</total>
-      </order>
-      <order>
-        <id>10460</id>
-        <orderdate>1997-02-28T00:00:00</orderdate>
-        <total>176.10</total>
-      </order>
-      <order>
-        <id>10533</id>
-        <orderdate>1997-05-12T00:00:00</orderdate>
-        <total>2222.20</total>
-      </order>
-      <order>
-        <id>10561</id>
-        <orderdate>1997-06-06T00:00:00</orderdate>
-        <total>2844.50</total>
-      </order>
-      <order>
-        <id>10703</id>
-        <orderdate>1997-10-14T00:00:00</orderdate>
-        <total>2545.00</total>
-      </order>
-      <order>
-        <id>10762</id>
-        <orderdate>1997-12-02T00:00:00</orderdate>
-        <total>4337.00</total>
-      </order>
-      <order>
-        <id>10774</id>
-        <orderdate>1997-12-11T00:00:00</orderdate>
-        <total>868.75</total>
-      </order>
-      <order>
-        <id>10824</id>
-        <orderdate>1998-01-09T00:00:00</orderdate>
-        <total>250.80</total>
-      </order>
-      <order>
-        <id>10880</id>
-        <orderdate>1998-02-10T00:00:00</orderdate>
-        <total>1500.00</total>
-      </order>
-      <order>
-        <id>10902</id>
-        <orderdate>1998-02-23T00:00:00</orderdate>
-        <total>863.43</total>
-      </order>
-      <order>
-        <id>10955</id>
-        <orderdate>1998-03-17T00:00:00</orderdate>
-        <total>74.40</total>
-      </order>
-      <order>
-        <id>10977</id>
-        <orderdate>1998-03-26T00:00:00</orderdate>
-        <total>2233.00</total>
-      </order>
-      <order>
-        <id>10980</id>
-        <orderdate>1998-03-27T00:00:00</orderdate>
-        <total>248.00</total>
-      </order>
-      <order>
-        <id>10993</id>
-        <orderdate>1998-04-01T00:00:00</orderdate>
-        <total>4895.44</total>
-      </order>
-      <order>
-        <id>11001</id>
-        <orderdate>1998-04-06T00:00:00</orderdate>
-        <total>2769.00</total>
-      </order>
-      <order>
-        <id>11050</id>
-        <orderdate>1998-04-27T00:00:00</orderdate>
-        <total>810.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>FRANK</id>
-    <name>Frankenversand</name>
-    <address>Berliner Platz 43</address>
-    <city>München</city>
-    <postalcode>80805</postalcode>
-    <country>Germany</country>
-    <phone>089-0877310</phone>
-    <fax>089-0877451</fax>
-    <orders>
-      <order>
-        <id>10267</id>
-        <orderdate>1996-07-29T00:00:00</orderdate>
-        <total>3536.60</total>
-      </order>
-      <order>
-        <id>10337</id>
-        <orderdate>1996-10-24T00:00:00</orderdate>
-        <total>2467.00</total>
-      </order>
-      <order>
-        <id>10342</id>
-        <orderdate>1996-10-30T00:00:00</orderdate>
-        <total>1840.64</total>
-      </order>
-      <order>
-        <id>10396</id>
-        <orderdate>1996-12-27T00:00:00</orderdate>
-        <total>1903.80</total>
-      </order>
-      <order>
-        <id>10488</id>
-        <orderdate>1997-03-27T00:00:00</orderdate>
-        <total>1512.00</total>
-      </order>
-      <order>
-        <id>10560</id>
-        <orderdate>1997-06-06T00:00:00</orderdate>
-        <total>1072.42</total>
-      </order>
-      <order>
-        <id>10623</id>
-        <orderdate>1997-08-07T00:00:00</orderdate>
-        <total>1336.95</total>
-      </order>
-      <order>
-        <id>10653</id>
-        <orderdate>1997-09-02T00:00:00</orderdate>
-        <total>1083.15</total>
-      </order>
-      <order>
-        <id>10670</id>
-        <orderdate>1997-09-16T00:00:00</orderdate>
-        <total>2301.75</total>
-      </order>
-      <order>
-        <id>10675</id>
-        <orderdate>1997-09-19T00:00:00</orderdate>
-        <total>1423.00</total>
-      </order>
-      <order>
-        <id>10717</id>
-        <orderdate>1997-10-24T00:00:00</orderdate>
-        <total>1270.75</total>
-      </order>
-      <order>
-        <id>10791</id>
-        <orderdate>1997-12-23T00:00:00</orderdate>
-        <total>1829.76</total>
-      </order>
-      <order>
-        <id>10859</id>
-        <orderdate>1998-01-29T00:00:00</orderdate>
-        <total>1078.69</total>
-      </order>
-      <order>
-        <id>10929</id>
-        <orderdate>1998-03-05T00:00:00</orderdate>
-        <total>1174.75</total>
-      </order>
-      <order>
-        <id>11012</id>
-        <orderdate>1998-04-09T00:00:00</orderdate>
-        <total>2825.30</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>FRANR</id>
-    <name>France restauration</name>
-    <address>54, rue Royale</address>
-    <city>Nantes</city>
-    <postalcode>44000</postalcode>
-    <country>France</country>
-    <phone>40.32.21.21</phone>
-    <fax>40.32.21.20</fax>
-    <orders>
-      <order>
-        <id>10671</id>
-        <orderdate>1997-09-17T00:00:00</orderdate>
-        <total>920.10</total>
-      </order>
-      <order>
-        <id>10860</id>
-        <orderdate>1998-01-29T00:00:00</orderdate>
-        <total>519.00</total>
-      </order>
-      <order>
-        <id>10971</id>
-        <orderdate>1998-03-24T00:00:00</orderdate>
-        <total>1733.06</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>FRANS</id>
-    <name>Franchi S.p.A.</name>
-    <address>Via Monte Bianco 34</address>
-    <city>Torino</city>
-    <postalcode>10100</postalcode>
-    <country>Italy</country>
-    <phone>011-4988260</phone>
-    <fax>011-4988261</fax>
-    <orders>
-      <order>
-        <id>10422</id>
-        <orderdate>1997-01-22T00:00:00</orderdate>
-        <total>49.80</total>
-      </order>
-      <order>
-        <id>10710</id>
-        <orderdate>1997-10-20T00:00:00</orderdate>
-        <total>93.50</total>
-      </order>
-      <order>
-        <id>10753</id>
-        <orderdate>1997-11-25T00:00:00</orderdate>
-        <total>88.00</total>
-      </order>
-      <order>
-        <id>10807</id>
-        <orderdate>1997-12-31T00:00:00</orderdate>
-        <total>18.40</total>
-      </order>
-      <order>
-        <id>11026</id>
-        <orderdate>1998-04-15T00:00:00</orderdate>
-        <total>1030.00</total>
-      </order>
-      <order>
-        <id>11060</id>
-        <orderdate>1998-04-30T00:00:00</orderdate>
-        <total>266.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>FURIB</id>
-    <name>Furia Bacalhau e Frutos do Mar</name>
-    <address>Jardim das rosas n. 32</address>
-    <city>Lisboa</city>
-    <postalcode>1675</postalcode>
-    <country>Portugal</country>
-    <phone>(1) 354-2534</phone>
-    <fax>(1) 354-2535</fax>
-    <orders>
-      <order>
-        <id>10328</id>
-        <orderdate>1996-10-14T00:00:00</orderdate>
-        <total>1168.00</total>
-      </order>
-      <order>
-        <id>10352</id>
-        <orderdate>1996-11-12T00:00:00</orderdate>
-        <total>136.30</total>
-      </order>
-      <order>
-        <id>10464</id>
-        <orderdate>1997-03-04T00:00:00</orderdate>
-        <total>1609.28</total>
-      </order>
-      <order>
-        <id>10491</id>
-        <orderdate>1997-03-31T00:00:00</orderdate>
-        <total>259.50</total>
-      </order>
-      <order>
-        <id>10551</id>
-        <orderdate>1997-05-28T00:00:00</orderdate>
-        <total>1677.30</total>
-      </order>
-      <order>
-        <id>10604</id>
-        <orderdate>1997-07-18T00:00:00</orderdate>
-        <total>230.85</total>
-      </order>
-      <order>
-        <id>10664</id>
-        <orderdate>1997-09-10T00:00:00</orderdate>
-        <total>1288.39</total>
-      </order>
-      <order>
-        <id>10963</id>
-        <orderdate>1998-03-19T00:00:00</orderdate>
-        <total>57.80</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>GALED</id>
-    <name>Galería del gastrónomo</name>
-    <address>Rambla de Cataluña, 23</address>
-    <city>Barcelona</city>
-    <postalcode>08022</postalcode>
-    <country>Spain</country>
-    <phone>(93) 203 4560</phone>
-    <fax>(93) 203 4561</fax>
-    <orders>
-      <order>
-        <id>10366</id>
-        <orderdate>1996-11-28T00:00:00</orderdate>
-        <total>136.00</total>
-      </order>
-      <order>
-        <id>10426</id>
-        <orderdate>1997-01-27T00:00:00</orderdate>
-        <total>338.20</total>
-      </order>
-      <order>
-        <id>10568</id>
-        <orderdate>1997-06-13T00:00:00</orderdate>
-        <total>155.00</total>
-      </order>
-      <order>
-        <id>10887</id>
-        <orderdate>1998-02-13T00:00:00</orderdate>
-        <total>70.00</total>
-      </order>
-      <order>
-        <id>10928</id>
-        <orderdate>1998-03-05T00:00:00</orderdate>
-        <total>137.50</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>GODOS</id>
-    <name>Godos Cocina Típica</name>
-    <address>C/ Romero, 33</address>
-    <city>Sevilla</city>
-    <postalcode>41101</postalcode>
-    <country>Spain</country>
-    <phone>(95) 555 82 82</phone>
-    <orders>
-      <order>
-        <id>10303</id>
-        <orderdate>1996-09-11T00:00:00</orderdate>
-        <total>1117.80</total>
-      </order>
-      <order>
-        <id>10550</id>
-        <orderdate>1997-05-28T00:00:00</orderdate>
-        <total>683.30</total>
-      </order>
-      <order>
-        <id>10629</id>
-        <orderdate>1997-08-12T00:00:00</orderdate>
-        <total>2775.05</total>
-      </order>
-      <order>
-        <id>10872</id>
-        <orderdate>1998-02-05T00:00:00</orderdate>
-        <total>2058.46</total>
-      </order>
-      <order>
-        <id>10874</id>
-        <orderdate>1998-02-06T00:00:00</orderdate>
-        <total>310.00</total>
-      </order>
-      <order>
-        <id>10888</id>
-        <orderdate>1998-02-16T00:00:00</orderdate>
-        <total>605.00</total>
-      </order>
-      <order>
-        <id>10911</id>
-        <orderdate>1998-02-26T00:00:00</orderdate>
-        <total>858.00</total>
-      </order>
-      <order>
-        <id>10948</id>
-        <orderdate>1998-03-13T00:00:00</orderdate>
-        <total>2362.25</total>
-      </order>
-      <order>
-        <id>11009</id>
-        <orderdate>1998-04-08T00:00:00</orderdate>
-        <total>616.50</total>
-      </order>
-      <order>
-        <id>11037</id>
-        <orderdate>1998-04-21T00:00:00</orderdate>
-        <total>60.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>GOURL</id>
-    <name>Gourmet Lanchonetes</name>
-    <address>Av.Brasil, 442</address>
-    <city>Campinas</city>
-    <region>SP</region>
-    <postalcode>04876-786</postalcode>
-    <country>Brazil</country>
-    <phone>(11) 555-9482</phone>
-    <orders>
-      <order>
-        <id>10423</id>
-        <orderdate>1997-01-23T00:00:00</orderdate>
-        <total>1020.00</total>
-      </order>
-      <order>
-        <id>10652</id>
-        <orderdate>1997-09-01T00:00:00</orderdate>
-        <total>318.84</total>
-      </order>
-      <order>
-        <id>10685</id>
-        <orderdate>1997-09-29T00:00:00</orderdate>
-        <total>801.10</total>
-      </order>
-      <order>
-        <id>10709</id>
-        <orderdate>1997-10-17T00:00:00</orderdate>
-        <total>3424.00</total>
-      </order>
-      <order>
-        <id>10734</id>
-        <orderdate>1997-11-07T00:00:00</orderdate>
-        <total>1498.35</total>
-      </order>
-      <order>
-        <id>10777</id>
-        <orderdate>1997-12-15T00:00:00</orderdate>
-        <total>224.00</total>
-      </order>
-      <order>
-        <id>10790</id>
-        <orderdate>1997-12-22T00:00:00</orderdate>
-        <total>722.50</total>
-      </order>
-      <order>
-        <id>10959</id>
-        <orderdate>1998-03-18T00:00:00</orderdate>
-        <total>131.75</total>
-      </order>
-      <order>
-        <id>11049</id>
-        <orderdate>1998-04-24T00:00:00</orderdate>
-        <total>273.60</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>GREAL</id>
-    <name>Great Lakes Food Market</name>
-    <address>2732 Baker Blvd.</address>
-    <city>Eugene</city>
-    <region>OR</region>
-    <postalcode>97403</postalcode>
-    <country>USA</country>
-    <phone>(503) 555-7555</phone>
-    <orders>
-      <order>
-        <id>10528</id>
-        <orderdate>1997-05-06T00:00:00</orderdate>
-        <total>392.20</total>
-      </order>
-      <order>
-        <id>10589</id>
-        <orderdate>1997-07-04T00:00:00</orderdate>
-        <total>72.00</total>
-      </order>
-      <order>
-        <id>10616</id>
-        <orderdate>1997-07-31T00:00:00</orderdate>
-        <total>4807.00</total>
-      </order>
-      <order>
-        <id>10617</id>
-        <orderdate>1997-07-31T00:00:00</orderdate>
-        <total>1402.50</total>
-      </order>
-      <order>
-        <id>10656</id>
-        <orderdate>1997-09-04T00:00:00</orderdate>
-        <total>604.22</total>
-      </order>
-      <order>
-        <id>10681</id>
-        <orderdate>1997-09-25T00:00:00</orderdate>
-        <total>1287.40</total>
-      </order>
-      <order>
-        <id>10816</id>
-        <orderdate>1998-01-06T00:00:00</orderdate>
-        <total>8446.45</total>
-      </order>
-      <order>
-        <id>10936</id>
-        <orderdate>1998-03-09T00:00:00</orderdate>
-        <total>456.00</total>
-      </order>
-      <order>
-        <id>11006</id>
-        <orderdate>1998-04-07T00:00:00</orderdate>
-        <total>329.68</total>
-      </order>
-      <order>
-        <id>11040</id>
-        <orderdate>1998-04-22T00:00:00</orderdate>
-        <total>200.00</total>
-      </order>
-      <order>
-        <id>11061</id>
-        <orderdate>1998-04-30T00:00:00</orderdate>
-        <total>510.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>GROSR</id>
-    <name>GROSELLA-Restaurante</name>
-    <address>5ª Ave.Los Palos Grandes</address>
-    <city>Caracas</city>
-    <region>DF</region>
-    <postalcode>1081</postalcode>
-    <country>Venezuela</country>
-    <phone>(2) 283-2951</phone>
-    <fax>(2) 283-3397</fax>
-    <orders>
-      <order>
-        <id>10268</id>
-        <orderdate>1996-07-30T00:00:00</orderdate>
-        <total>1101.20</total>
-      </order>
-      <order>
-        <id>10785</id>
-        <orderdate>1997-12-18T00:00:00</orderdate>
-        <total>387.50</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>HANAR</id>
-    <name>Hanari Carnes</name>
-    <address>Rua do Paço, 67</address>
-    <city>Rio de Janeiro</city>
-    <region>RJ</region>
-    <postalcode>05454-876</postalcode>
-    <country>Brazil</country>
-    <phone>(21) 555-0091</phone>
-    <fax>(21) 555-8765</fax>
-    <orders>
-      <order>
-        <id>10250</id>
-        <orderdate>1996-07-08T00:00:00</orderdate>
-        <total>1552.60</total>
-      </order>
-      <order>
-        <id>10253</id>
-        <orderdate>1996-07-10T00:00:00</orderdate>
-        <total>1444.80</total>
-      </order>
-      <order>
-        <id>10541</id>
-        <orderdate>1997-05-19T00:00:00</orderdate>
-        <total>1946.52</total>
-      </order>
-      <order>
-        <id>10645</id>
-        <orderdate>1997-08-26T00:00:00</orderdate>
-        <total>1535.00</total>
-      </order>
-      <order>
-        <id>10690</id>
-        <orderdate>1997-10-02T00:00:00</orderdate>
-        <total>862.50</total>
-      </order>
-      <order>
-        <id>10770</id>
-        <orderdate>1997-12-09T00:00:00</orderdate>
-        <total>236.25</total>
-      </order>
-      <order>
-        <id>10783</id>
-        <orderdate>1997-12-18T00:00:00</orderdate>
-        <total>1442.50</total>
-      </order>
-      <order>
-        <id>10886</id>
-        <orderdate>1998-02-13T00:00:00</orderdate>
-        <total>3127.50</total>
-      </order>
-      <order>
-        <id>10903</id>
-        <orderdate>1998-02-24T00:00:00</orderdate>
-        <total>932.05</total>
-      </order>
-      <order>
-        <id>10922</id>
-        <orderdate>1998-03-03T00:00:00</orderdate>
-        <total>742.50</total>
-      </order>
-      <order>
-        <id>10925</id>
-        <orderdate>1998-03-04T00:00:00</orderdate>
-        <total>475.15</total>
-      </order>
-      <order>
-        <id>10981</id>
-        <orderdate>1998-03-27T00:00:00</orderdate>
-        <total>15810.00</total>
-      </order>
-      <order>
-        <id>11022</id>
-        <orderdate>1998-04-14T00:00:00</orderdate>
-        <total>1402.00</total>
-      </order>
-      <order>
-        <id>11052</id>
-        <orderdate>1998-04-27T00:00:00</orderdate>
-        <total>1332.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>HILAA</id>
-    <name>HILARIÓN-Abastos</name>
-    <address>Carrera 22 con Ave.Carlos Soublette #8-35</address>
-    <city> San Cristóbal</city>
-    <region>Táchira</region>
-    <postalcode>5022</postalcode>
-    <country>Venezuela</country>
-    <phone>(5) 555-1340</phone>
-    <fax>(5) 555-1948</fax>
-    <orders>
-      <order>
-        <id>10257</id>
-        <orderdate>1996-07-16T00:00:00</orderdate>
-        <total>1119.90</total>
-      </order>
-      <order>
-        <id>10395</id>
-        <orderdate>1996-12-26T00:00:00</orderdate>
-        <total>2122.92</total>
-      </order>
-      <order>
-        <id>10476</id>
-        <orderdate>1997-03-17T00:00:00</orderdate>
-        <total>180.48</total>
-      </order>
-      <order>
-        <id>10486</id>
-        <orderdate>1997-03-26T00:00:00</orderdate>
-        <total>1272.00</total>
-      </order>
-      <order>
-        <id>10490</id>
-        <orderdate>1997-03-31T00:00:00</orderdate>
-        <total>3163.20</total>
-      </order>
-      <order>
-        <id>10498</id>
-        <orderdate>1997-04-07T00:00:00</orderdate>
-        <total>575.00</total>
-      </order>
-      <order>
-        <id>10552</id>
-        <orderdate>1997-05-29T00:00:00</orderdate>
-        <total>880.50</total>
-      </order>
-      <order>
-        <id>10601</id>
-        <orderdate>1997-07-16T00:00:00</orderdate>
-        <total>2285.00</total>
-      </order>
-      <order>
-        <id>10613</id>
-        <orderdate>1997-07-29T00:00:00</orderdate>
-        <total>353.20</total>
-      </order>
-      <order>
-        <id>10641</id>
-        <orderdate>1997-08-22T00:00:00</orderdate>
-        <total>2054.00</total>
-      </order>
-      <order>
-        <id>10705</id>
-        <orderdate>1997-10-15T00:00:00</orderdate>
-        <total>378.00</total>
-      </order>
-      <order>
-        <id>10796</id>
-        <orderdate>1997-12-25T00:00:00</orderdate>
-        <total>2341.36</total>
-      </order>
-      <order>
-        <id>10863</id>
-        <orderdate>1998-02-02T00:00:00</orderdate>
-        <total>441.15</total>
-      </order>
-      <order>
-        <id>10901</id>
-        <orderdate>1998-02-23T00:00:00</orderdate>
-        <total>934.50</total>
-      </order>
-      <order>
-        <id>10957</id>
-        <orderdate>1998-03-18T00:00:00</orderdate>
-        <total>1762.70</total>
-      </order>
-      <order>
-        <id>10960</id>
-        <orderdate>1998-03-19T00:00:00</orderdate>
-        <total>265.35</total>
-      </order>
-      <order>
-        <id>10976</id>
-        <orderdate>1998-03-25T00:00:00</orderdate>
-        <total>912.00</total>
-      </order>
-      <order>
-        <id>11055</id>
-        <orderdate>1998-04-28T00:00:00</orderdate>
-        <total>1727.50</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>HUNGC</id>
-    <name>Hungry Coyote Import Store</name>
-    <address>City Center Plaza, 516 Main St.</address>
-    <city>Elgin</city>
-    <region>OR</region>
-    <postalcode>97827</postalcode>
-    <country>USA</country>
-    <phone>(503) 555-6874</phone>
-    <fax>(503) 555-2376</fax>
-    <orders>
-      <order>
-        <id>10375</id>
-        <orderdate>1996-12-06T00:00:00</orderdate>
-        <total>338.00</total>
-      </order>
-      <order>
-        <id>10394</id>
-        <orderdate>1996-12-25T00:00:00</orderdate>
-        <total>442.00</total>
-      </order>
-      <order>
-        <id>10415</id>
-        <orderdate>1997-01-15T00:00:00</orderdate>
-        <total>102.40</total>
-      </order>
-      <order>
-        <id>10600</id>
-        <orderdate>1997-07-16T00:00:00</orderdate>
-        <total>479.80</total>
-      </order>
-      <order>
-        <id>10660</id>
-        <orderdate>1997-09-08T00:00:00</orderdate>
-        <total>1701.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>HUNGO</id>
-    <name>Hungry Owl All-Night Grocers</name>
-    <address>8 Johnstown Road</address>
-    <city>Cork</city>
-    <region>Co.Cork</region>
-    <country>Ireland</country>
-    <phone>2967 542</phone>
-    <fax>2967 3333</fax>
-    <orders>
-      <order>
-        <id>10298</id>
-        <orderdate>1996-09-05T00:00:00</orderdate>
-        <total>2645.00</total>
-      </order>
-      <order>
-        <id>10309</id>
-        <orderdate>1996-09-19T00:00:00</orderdate>
-        <total>1762.00</total>
-      </order>
-      <order>
-        <id>10335</id>
-        <orderdate>1996-10-22T00:00:00</orderdate>
-        <total>2036.16</total>
-      </order>
-      <order>
-        <id>10373</id>
-        <orderdate>1996-12-05T00:00:00</orderdate>
-        <total>1366.40</total>
-      </order>
-      <order>
-        <id>10380</id>
-        <orderdate>1996-12-12T00:00:00</orderdate>
-        <total>1313.82</total>
-      </order>
-      <order>
-        <id>10429</id>
-        <orderdate>1997-01-29T00:00:00</orderdate>
-        <total>1441.38</total>
-      </order>
-      <order>
-        <id>10503</id>
-        <orderdate>1997-04-11T00:00:00</orderdate>
-        <total>2048.50</total>
-      </order>
-      <order>
-        <id>10516</id>
-        <orderdate>1997-04-24T00:00:00</orderdate>
-        <total>2381.05</total>
-      </order>
-      <order>
-        <id>10567</id>
-        <orderdate>1997-06-12T00:00:00</orderdate>
-        <total>2519.00</total>
-      </order>
-      <order>
-        <id>10646</id>
-        <orderdate>1997-08-27T00:00:00</orderdate>
-        <total>1446.00</total>
-      </order>
-      <order>
-        <id>10661</id>
-        <orderdate>1997-09-09T00:00:00</orderdate>
-        <total>562.60</total>
-      </order>
-      <order>
-        <id>10687</id>
-        <orderdate>1997-09-30T00:00:00</orderdate>
-        <total>4960.90</total>
-      </order>
-      <order>
-        <id>10701</id>
-        <orderdate>1997-10-13T00:00:00</orderdate>
-        <total>2864.50</total>
-      </order>
-      <order>
-        <id>10712</id>
-        <orderdate>1997-10-21T00:00:00</orderdate>
-        <total>1233.48</total>
-      </order>
-      <order>
-        <id>10736</id>
-        <orderdate>1997-11-11T00:00:00</orderdate>
-        <total>997.00</total>
-      </order>
-      <order>
-        <id>10897</id>
-        <orderdate>1998-02-19T00:00:00</orderdate>
-        <total>10835.24</total>
-      </order>
-      <order>
-        <id>10912</id>
-        <orderdate>1998-02-26T00:00:00</orderdate>
-        <total>6200.55</total>
-      </order>
-      <order>
-        <id>10985</id>
-        <orderdate>1998-03-30T00:00:00</orderdate>
-        <total>2023.38</total>
-      </order>
-      <order>
-        <id>11063</id>
-        <orderdate>1998-04-30T00:00:00</orderdate>
-        <total>1342.95</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>ISLAT</id>
-    <name>Island Trading</name>
-    <address>Garden House, Crowther Way</address>
-    <city>Cowes</city>
-    <region>Isle of Wight</region>
-    <postalcode>PO31 7PJ</postalcode>
-    <country>UK</country>
-    <phone>(198) 555-8888</phone>
-    <orders>
-      <order>
-        <id>10315</id>
-        <orderdate>1996-09-26T00:00:00</orderdate>
-        <total>516.80</total>
-      </order>
-      <order>
-        <id>10318</id>
-        <orderdate>1996-10-01T00:00:00</orderdate>
-        <total>240.40</total>
-      </order>
-      <order>
-        <id>10321</id>
-        <orderdate>1996-10-03T00:00:00</orderdate>
-        <total>144.00</total>
-      </order>
-      <order>
-        <id>10473</id>
-        <orderdate>1997-03-13T00:00:00</orderdate>
-        <total>230.40</total>
-      </order>
-      <order>
-        <id>10621</id>
-        <orderdate>1997-08-05T00:00:00</orderdate>
-        <total>758.50</total>
-      </order>
-      <order>
-        <id>10674</id>
-        <orderdate>1997-09-18T00:00:00</orderdate>
-        <total>45.00</total>
-      </order>
-      <order>
-        <id>10749</id>
-        <orderdate>1997-11-20T00:00:00</orderdate>
-        <total>1080.00</total>
-      </order>
-      <order>
-        <id>10798</id>
-        <orderdate>1997-12-26T00:00:00</orderdate>
-        <total>446.60</total>
-      </order>
-      <order>
-        <id>10829</id>
-        <orderdate>1998-01-13T00:00:00</orderdate>
-        <total>1764.00</total>
-      </order>
-      <order>
-        <id>10933</id>
-        <orderdate>1998-03-06T00:00:00</orderdate>
-        <total>920.60</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>KOENE</id>
-    <name>Königlich Essen</name>
-    <address>Maubelstr. 90</address>
-    <city>Brandenburg</city>
-    <postalcode>14776</postalcode>
-    <country>Germany</country>
-    <phone>0555-09876</phone>
-    <orders>
-      <order>
-        <id>10323</id>
-        <orderdate>1996-10-07T00:00:00</orderdate>
-        <total>164.40</total>
-      </order>
-      <order>
-        <id>10325</id>
-        <orderdate>1996-10-09T00:00:00</orderdate>
-        <total>1497.00</total>
-      </order>
-      <order>
-        <id>10456</id>
-        <orderdate>1997-02-25T00:00:00</orderdate>
-        <total>557.60</total>
-      </order>
-      <order>
-        <id>10457</id>
-        <orderdate>1997-02-25T00:00:00</orderdate>
-        <total>1584.00</total>
-      </order>
-      <order>
-        <id>10468</id>
-        <orderdate>1997-03-07T00:00:00</orderdate>
-        <total>717.60</total>
-      </order>
-      <order>
-        <id>10506</id>
-        <orderdate>1997-04-15T00:00:00</orderdate>
-        <total>415.80</total>
-      </order>
-      <order>
-        <id>10542</id>
-        <orderdate>1997-05-20T00:00:00</orderdate>
-        <total>469.11</total>
-      </order>
-      <order>
-        <id>10630</id>
-        <orderdate>1997-08-13T00:00:00</orderdate>
-        <total>903.60</total>
-      </order>
-      <order>
-        <id>10718</id>
-        <orderdate>1997-10-27T00:00:00</orderdate>
-        <total>3463.00</total>
-      </order>
-      <order>
-        <id>10799</id>
-        <orderdate>1997-12-26T00:00:00</orderdate>
-        <total>1553.50</total>
-      </order>
-      <order>
-        <id>10817</id>
-        <orderdate>1998-01-06T00:00:00</orderdate>
-        <total>10952.84</total>
-      </order>
-      <order>
-        <id>10849</id>
-        <orderdate>1998-01-23T00:00:00</orderdate>
-        <total>967.82</total>
-      </order>
-      <order>
-        <id>10893</id>
-        <orderdate>1998-02-18T00:00:00</orderdate>
-        <total>5502.11</total>
-      </order>
-      <order>
-        <id>11028</id>
-        <orderdate>1998-04-16T00:00:00</orderdate>
-        <total>2160.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>LACOR</id>
-    <name>La corne d'abondance</name>
-    <address>67, avenue de l'Europe</address>
-    <city>Versailles</city>
-    <postalcode>78000</postalcode>
-    <country>France</country>
-    <phone>30.59.84.10</phone>
-    <fax>30.59.85.11</fax>
-    <orders>
-      <order>
-        <id>10858</id>
-        <orderdate>1998-01-29T00:00:00</orderdate>
-        <total>649.00</total>
-      </order>
-      <order>
-        <id>10927</id>
-        <orderdate>1998-03-05T00:00:00</orderdate>
-        <total>800.00</total>
-      </order>
-      <order>
-        <id>10972</id>
-        <orderdate>1998-03-24T00:00:00</orderdate>
-        <total>251.50</total>
-      </order>
-      <order>
-        <id>10973</id>
-        <orderdate>1998-03-24T00:00:00</orderdate>
-        <total>291.55</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>LAMAI</id>
-    <name>La maison d'Asie</name>
-    <address>1 rue Alsace-Lorraine</address>
-    <city>Toulouse</city>
-    <postalcode>31000</postalcode>
-    <country>France</country>
-    <phone>61.77.61.10</phone>
-    <fax>61.77.61.11</fax>
-    <orders>
-      <order>
-        <id>10350</id>
-        <orderdate>1996-11-11T00:00:00</orderdate>
-        <total>642.06</total>
-      </order>
-      <order>
-        <id>10358</id>
-        <orderdate>1996-11-20T00:00:00</orderdate>
-        <total>429.40</total>
-      </order>
-      <order>
-        <id>10371</id>
-        <orderdate>1996-12-03T00:00:00</orderdate>
-        <total>72.96</total>
-      </order>
-      <order>
-        <id>10413</id>
-        <orderdate>1997-01-14T00:00:00</orderdate>
-        <total>2123.20</total>
-      </order>
-      <order>
-        <id>10425</id>
-        <orderdate>1997-01-24T00:00:00</orderdate>
-        <total>360.00</total>
-      </order>
-      <order>
-        <id>10454</id>
-        <orderdate>1997-02-21T00:00:00</orderdate>
-        <total>331.20</total>
-      </order>
-      <order>
-        <id>10493</id>
-        <orderdate>1997-04-02T00:00:00</orderdate>
-        <total>608.40</total>
-      </order>
-      <order>
-        <id>10500</id>
-        <orderdate>1997-04-09T00:00:00</orderdate>
-        <total>523.26</total>
-      </order>
-      <order>
-        <id>10610</id>
-        <orderdate>1997-07-25T00:00:00</orderdate>
-        <total>299.25</total>
-      </order>
-      <order>
-        <id>10631</id>
-        <orderdate>1997-08-14T00:00:00</orderdate>
-        <total>55.80</total>
-      </order>
-      <order>
-        <id>10787</id>
-        <orderdate>1997-12-19T00:00:00</orderdate>
-        <total>2622.76</total>
-      </order>
-      <order>
-        <id>10832</id>
-        <orderdate>1998-01-14T00:00:00</orderdate>
-        <total>475.11</total>
-      </order>
-      <order>
-        <id>10923</id>
-        <orderdate>1998-03-03T00:00:00</orderdate>
-        <total>748.80</total>
-      </order>
-      <order>
-        <id>11051</id>
-        <orderdate>1998-04-27T00:00:00</orderdate>
-        <total>36.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>LAUGB</id>
-    <name>Laughing Bacchus Wine Cellars</name>
-    <address>1900 Oak St.</address>
-    <city>Vancouver</city>
-    <region>BC</region>
-    <postalcode>V3F 2K1</postalcode>
-    <country>Canada</country>
-    <phone>(604) 555-3392</phone>
-    <fax>(604) 555-7293</fax>
-    <orders>
-      <order>
-        <id>10495</id>
-        <orderdate>1997-04-03T00:00:00</orderdate>
-        <total>278.00</total>
-      </order>
-      <order>
-        <id>10620</id>
-        <orderdate>1997-08-05T00:00:00</orderdate>
-        <total>57.50</total>
-      </order>
-      <order>
-        <id>10810</id>
-        <orderdate>1998-01-01T00:00:00</orderdate>
-        <total>187.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>LAZYK</id>
-    <name>Lazy K Kountry Store</name>
-    <address>12 Orchestra Terrace</address>
-    <city>Walla Walla</city>
-    <region>WA</region>
-    <postalcode>99362</postalcode>
-    <country>USA</country>
-    <phone>(509) 555-7969</phone>
-    <fax>(509) 555-6221</fax>
-    <orders>
-      <order>
-        <id>10482</id>
-        <orderdate>1997-03-21T00:00:00</orderdate>
-        <total>147.00</total>
-      </order>
-      <order>
-        <id>10545</id>
-        <orderdate>1997-05-22T00:00:00</orderdate>
-        <total>210.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>LEHMS</id>
-    <name>Lehmanns Marktstand</name>
-    <address>Magazinweg 7</address>
-    <city>Frankfurt a.M. </city>
-    <postalcode>60528</postalcode>
-    <country>Germany</country>
-    <phone>069-0245984</phone>
-    <fax>069-0245874</fax>
-    <orders>
-      <order>
-        <id>10279</id>
-        <orderdate>1996-08-13T00:00:00</orderdate>
-        <total>351.00</total>
-      </order>
-      <order>
-        <id>10284</id>
-        <orderdate>1996-08-19T00:00:00</orderdate>
-        <total>1170.38</total>
-      </order>
-      <order>
-        <id>10343</id>
-        <orderdate>1996-10-31T00:00:00</orderdate>
-        <total>1584.00</total>
-      </order>
-      <order>
-        <id>10497</id>
-        <orderdate>1997-04-04T00:00:00</orderdate>
-        <total>1380.60</total>
-      </order>
-      <order>
-        <id>10522</id>
-        <orderdate>1997-04-30T00:00:00</orderdate>
-        <total>2318.24</total>
-      </order>
-      <order>
-        <id>10534</id>
-        <orderdate>1997-05-12T00:00:00</orderdate>
-        <total>465.70</total>
-      </order>
-      <order>
-        <id>10536</id>
-        <orderdate>1997-05-14T00:00:00</orderdate>
-        <total>1645.00</total>
-      </order>
-      <order>
-        <id>10557</id>
-        <orderdate>1997-06-03T00:00:00</orderdate>
-        <total>1152.50</total>
-      </order>
-      <order>
-        <id>10592</id>
-        <orderdate>1997-07-08T00:00:00</orderdate>
-        <total>516.47</total>
-      </order>
-      <order>
-        <id>10593</id>
-        <orderdate>1997-07-09T00:00:00</orderdate>
-        <total>1994.40</total>
-      </order>
-      <order>
-        <id>10772</id>
-        <orderdate>1997-12-10T00:00:00</orderdate>
-        <total>3603.22</total>
-      </order>
-      <order>
-        <id>10862</id>
-        <orderdate>1998-01-30T00:00:00</orderdate>
-        <total>581.00</total>
-      </order>
-      <order>
-        <id>10891</id>
-        <orderdate>1998-02-17T00:00:00</orderdate>
-        <total>368.93</total>
-      </order>
-      <order>
-        <id>10934</id>
-        <orderdate>1998-03-09T00:00:00</orderdate>
-        <total>500.00</total>
-      </order>
-      <order>
-        <id>11070</id>
-        <orderdate>1998-05-05T00:00:00</orderdate>
-        <total>1629.98</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>LETSS</id>
-    <name>Let's Stop N Shop</name>
-    <address>87 Polk St. Suite 5</address>
-    <city>San Francisco</city>
-    <region>CA</region>
-    <postalcode>94117</postalcode>
-    <country>USA</country>
-    <phone>(415) 555-5938</phone>
-    <orders>
-      <order>
-        <id>10579</id>
-        <orderdate>1997-06-25T00:00:00</orderdate>
-        <total>317.75</total>
-      </order>
-      <order>
-        <id>10719</id>
-        <orderdate>1997-10-27T00:00:00</orderdate>
-        <total>844.25</total>
-      </order>
-      <order>
-        <id>10735</id>
-        <orderdate>1997-11-10T00:00:00</orderdate>
-        <total>536.40</total>
-      </order>
-      <order>
-        <id>10884</id>
-        <orderdate>1998-02-12T00:00:00</orderdate>
-        <total>1378.07</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>LILAS</id>
-    <name>LILA-Supermercado</name>
-    <address>Carrera 52 con Ave.Bolívar #65-98 Llano Largo</address>
-    <city>Barquisimeto</city>
-    <region>Lara</region>
-    <postalcode>3508</postalcode>
-    <country>Venezuela</country>
-    <phone>(9) 331-6954</phone>
-    <fax>(9) 331-7256</fax>
-    <orders>
-      <order>
-        <id>10283</id>
-        <orderdate>1996-08-16T00:00:00</orderdate>
-        <total>1414.80</total>
-      </order>
-      <order>
-        <id>10296</id>
-        <orderdate>1996-09-03T00:00:00</orderdate>
-        <total>1050.60</total>
-      </order>
-      <order>
-        <id>10330</id>
-        <orderdate>1996-10-16T00:00:00</orderdate>
-        <total>1649.00</total>
-      </order>
-      <order>
-        <id>10357</id>
-        <orderdate>1996-11-19T00:00:00</orderdate>
-        <total>1167.68</total>
-      </order>
-      <order>
-        <id>10381</id>
-        <orderdate>1996-12-12T00:00:00</orderdate>
-        <total>112.00</total>
-      </order>
-      <order>
-        <id>10461</id>
-        <orderdate>1997-02-28T00:00:00</orderdate>
-        <total>1538.70</total>
-      </order>
-      <order>
-        <id>10499</id>
-        <orderdate>1997-04-08T00:00:00</orderdate>
-        <total>1412.00</total>
-      </order>
-      <order>
-        <id>10543</id>
-        <orderdate>1997-05-21T00:00:00</orderdate>
-        <total>1504.50</total>
-      </order>
-      <order>
-        <id>10780</id>
-        <orderdate>1997-12-16T00:00:00</orderdate>
-        <total>720.00</total>
-      </order>
-      <order>
-        <id>10823</id>
-        <orderdate>1998-01-09T00:00:00</orderdate>
-        <total>2826.00</total>
-      </order>
-      <order>
-        <id>10899</id>
-        <orderdate>1998-02-20T00:00:00</orderdate>
-        <total>122.40</total>
-      </order>
-      <order>
-        <id>10997</id>
-        <orderdate>1998-04-03T00:00:00</orderdate>
-        <total>1885.00</total>
-      </order>
-      <order>
-        <id>11065</id>
-        <orderdate>1998-05-01T00:00:00</orderdate>
-        <total>189.42</total>
-      </order>
-      <order>
-        <id>11071</id>
-        <orderdate>1998-05-05T00:00:00</orderdate>
-        <total>484.50</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>LINOD</id>
-    <name>LINO-Delicateses</name>
-    <address>Ave. 5 de Mayo Porlamar</address>
-    <city>I.de Margarita</city>
-    <region>Nueva Esparta</region>
-    <postalcode>4980</postalcode>
-    <country>Venezuela</country>
-    <phone>(8) 34-56-12</phone>
-    <fax>(8) 34-93-93</fax>
-    <orders>
-      <order>
-        <id>10405</id>
-        <orderdate>1997-01-06T00:00:00</orderdate>
-        <total>400.00</total>
-      </order>
-      <order>
-        <id>10485</id>
-        <orderdate>1997-03-25T00:00:00</orderdate>
-        <total>1584.00</total>
-      </order>
-      <order>
-        <id>10638</id>
-        <orderdate>1997-08-20T00:00:00</orderdate>
-        <total>2720.05</total>
-      </order>
-      <order>
-        <id>10697</id>
-        <orderdate>1997-10-08T00:00:00</orderdate>
-        <total>805.42</total>
-      </order>
-      <order>
-        <id>10729</id>
-        <orderdate>1997-11-04T00:00:00</orderdate>
-        <total>1850.00</total>
-      </order>
-      <order>
-        <id>10811</id>
-        <orderdate>1998-01-02T00:00:00</orderdate>
-        <total>852.00</total>
-      </order>
-      <order>
-        <id>10838</id>
-        <orderdate>1998-01-19T00:00:00</orderdate>
-        <total>1938.38</total>
-      </order>
-      <order>
-        <id>10840</id>
-        <orderdate>1998-01-19T00:00:00</orderdate>
-        <total>211.20</total>
-      </order>
-      <order>
-        <id>10919</id>
-        <orderdate>1998-03-02T00:00:00</orderdate>
-        <total>1122.80</total>
-      </order>
-      <order>
-        <id>10954</id>
-        <orderdate>1998-03-17T00:00:00</orderdate>
-        <total>1659.54</total>
-      </order>
-      <order>
-        <id>11014</id>
-        <orderdate>1998-04-10T00:00:00</orderdate>
-        <total>243.18</total>
-      </order>
-      <order>
-        <id>11039</id>
-        <orderdate>1998-04-21T00:00:00</orderdate>
-        <total>3090.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>LONEP</id>
-    <name>Lonesome Pine Restaurant</name>
-    <address>89 Chiaroscuro Rd.</address>
-    <city>Portland</city>
-    <region>OR</region>
-    <postalcode>97219</postalcode>
-    <country>USA</country>
-    <phone>(503) 555-9573</phone>
-    <fax>(503) 555-9646</fax>
-    <orders>
-      <order>
-        <id>10307</id>
-        <orderdate>1996-09-17T00:00:00</orderdate>
-        <total>424.00</total>
-      </order>
-      <order>
-        <id>10317</id>
-        <orderdate>1996-09-30T00:00:00</orderdate>
-        <total>288.00</total>
-      </order>
-      <order>
-        <id>10544</id>
-        <orderdate>1997-05-21T00:00:00</orderdate>
-        <total>417.20</total>
-      </order>
-      <order>
-        <id>10662</id>
-        <orderdate>1997-09-09T00:00:00</orderdate>
-        <total>125.00</total>
-      </order>
-      <order>
-        <id>10665</id>
-        <orderdate>1997-09-11T00:00:00</orderdate>
-        <total>1295.00</total>
-      </order>
-      <order>
-        <id>10867</id>
-        <orderdate>1998-02-03T00:00:00</orderdate>
-        <total>98.40</total>
-      </order>
-      <order>
-        <id>10883</id>
-        <orderdate>1998-02-12T00:00:00</orderdate>
-        <total>36.00</total>
-      </order>
-      <order>
-        <id>11018</id>
-        <orderdate>1998-04-13T00:00:00</orderdate>
-        <total>1575.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>MAGAA</id>
-    <name>Magazzini Alimentari Riuniti</name>
-    <address>Via Ludovico il Moro 22</address>
-    <city>Bergamo</city>
-    <postalcode>24100</postalcode>
-    <country>Italy</country>
-    <phone>035-640230</phone>
-    <fax>035-640231</fax>
-    <orders>
-      <order>
-        <id>10275</id>
-        <orderdate>1996-08-07T00:00:00</orderdate>
-        <total>291.84</total>
-      </order>
-      <order>
-        <id>10300</id>
-        <orderdate>1996-09-09T00:00:00</orderdate>
-        <total>608.00</total>
-      </order>
-      <order>
-        <id>10404</id>
-        <orderdate>1997-01-03T00:00:00</orderdate>
-        <total>1591.25</total>
-      </order>
-      <order>
-        <id>10467</id>
-        <orderdate>1997-03-06T00:00:00</orderdate>
-        <total>235.20</total>
-      </order>
-      <order>
-        <id>10635</id>
-        <orderdate>1997-08-18T00:00:00</orderdate>
-        <total>1326.22</total>
-      </order>
-      <order>
-        <id>10754</id>
-        <orderdate>1997-11-25T00:00:00</orderdate>
-        <total>55.20</total>
-      </order>
-      <order>
-        <id>10784</id>
-        <orderdate>1997-12-18T00:00:00</orderdate>
-        <total>1488.00</total>
-      </order>
-      <order>
-        <id>10818</id>
-        <orderdate>1998-01-07T00:00:00</orderdate>
-        <total>833.00</total>
-      </order>
-      <order>
-        <id>10939</id>
-        <orderdate>1998-03-10T00:00:00</orderdate>
-        <total>637.50</total>
-      </order>
-      <order>
-        <id>10950</id>
-        <orderdate>1998-03-16T00:00:00</orderdate>
-        <total>110.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>MAISD</id>
-    <name>Maison Dewey</name>
-    <address>Rue Joseph-Bens 532</address>
-    <city>Bruxelles</city>
-    <postalcode>B-1180</postalcode>
-    <country>Belgium</country>
-    <phone>(02) 201 24 67</phone>
-    <fax>(02) 201 24 68</fax>
-    <orders>
-      <order>
-        <id>10529</id>
-        <orderdate>1997-05-07T00:00:00</orderdate>
-        <total>946.00</total>
-      </order>
-      <order>
-        <id>10649</id>
-        <orderdate>1997-08-28T00:00:00</orderdate>
-        <total>1434.00</total>
-      </order>
-      <order>
-        <id>10760</id>
-        <orderdate>1997-12-01T00:00:00</orderdate>
-        <total>2917.00</total>
-      </order>
-      <order>
-        <id>10892</id>
-        <orderdate>1998-02-17T00:00:00</orderdate>
-        <total>2090.00</total>
-      </order>
-      <order>
-        <id>10896</id>
-        <orderdate>1998-02-19T00:00:00</orderdate>
-        <total>750.50</total>
-      </order>
-      <order>
-        <id>10978</id>
-        <orderdate>1998-03-26T00:00:00</orderdate>
-        <total>1303.20</total>
-      </order>
-      <order>
-        <id>11004</id>
-        <orderdate>1998-04-07T00:00:00</orderdate>
-        <total>295.38</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>MEREP</id>
-    <name>Mère Paillarde</name>
-    <address>43 rue St. Laurent</address>
-    <city>Montréal</city>
-    <region>Québec</region>
-    <postalcode>H1J 1C3</postalcode>
-    <country>Canada</country>
-    <phone>(514) 555-8054</phone>
-    <fax>(514) 555-8055</fax>
-    <orders>
-      <order>
-        <id>10332</id>
-        <orderdate>1996-10-17T00:00:00</orderdate>
-        <total>1786.88</total>
-      </order>
-      <order>
-        <id>10339</id>
-        <orderdate>1996-10-28T00:00:00</orderdate>
-        <total>3354.00</total>
-      </order>
-      <order>
-        <id>10376</id>
-        <orderdate>1996-12-09T00:00:00</orderdate>
-        <total>399.00</total>
-      </order>
-      <order>
-        <id>10424</id>
-        <orderdate>1997-01-23T00:00:00</orderdate>
-        <total>9194.56</total>
-      </order>
-      <order>
-        <id>10439</id>
-        <orderdate>1997-02-07T00:00:00</orderdate>
-        <total>1078.00</total>
-      </order>
-      <order>
-        <id>10505</id>
-        <orderdate>1997-04-14T00:00:00</orderdate>
-        <total>147.90</total>
-      </order>
-      <order>
-        <id>10565</id>
-        <orderdate>1997-06-11T00:00:00</orderdate>
-        <total>639.90</total>
-      </order>
-      <order>
-        <id>10570</id>
-        <orderdate>1997-06-17T00:00:00</orderdate>
-        <total>2465.25</total>
-      </order>
-      <order>
-        <id>10590</id>
-        <orderdate>1997-07-07T00:00:00</orderdate>
-        <total>1101.00</total>
-      </order>
-      <order>
-        <id>10605</id>
-        <orderdate>1997-07-21T00:00:00</orderdate>
-        <total>4109.70</total>
-      </order>
-      <order>
-        <id>10618</id>
-        <orderdate>1997-08-01T00:00:00</orderdate>
-        <total>2697.50</total>
-      </order>
-      <order>
-        <id>10619</id>
-        <orderdate>1997-08-04T00:00:00</orderdate>
-        <total>1260.00</total>
-      </order>
-      <order>
-        <id>10724</id>
-        <orderdate>1997-10-30T00:00:00</orderdate>
-        <total>638.50</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>MORGK</id>
-    <name>Morgenstern Gesundkost</name>
-    <address>Heerstr. 22</address>
-    <city>Leipzig</city>
-    <postalcode>04179</postalcode>
-    <country>Germany</country>
-    <phone>0342-023176</phone>
-    <orders>
-      <order>
-        <id>10277</id>
-        <orderdate>1996-08-09T00:00:00</orderdate>
-        <total>1200.80</total>
-      </order>
-      <order>
-        <id>10575</id>
-        <orderdate>1997-06-20T00:00:00</orderdate>
-        <total>2147.40</total>
-      </order>
-      <order>
-        <id>10699</id>
-        <orderdate>1997-10-09T00:00:00</orderdate>
-        <total>114.00</total>
-      </order>
-      <order>
-        <id>10779</id>
-        <orderdate>1997-12-16T00:00:00</orderdate>
-        <total>1335.00</total>
-      </order>
-      <order>
-        <id>10945</id>
-        <orderdate>1998-03-12T00:00:00</orderdate>
-        <total>245.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>NORTS</id>
-    <name>North/South</name>
-    <address>South House, 300 Queensbridge</address>
-    <city>London</city>
-    <postalcode>SW7 1RZ</postalcode>
-    <country>UK</country>
-    <phone>(171) 555-7733</phone>
-    <fax>(171) 555-2530</fax>
-    <orders>
-      <order>
-        <id>10517</id>
-        <orderdate>1997-04-24T00:00:00</orderdate>
-        <total>352.00</total>
-      </order>
-      <order>
-        <id>10752</id>
-        <orderdate>1997-11-24T00:00:00</orderdate>
-        <total>252.00</total>
-      </order>
-      <order>
-        <id>11057</id>
-        <orderdate>1998-04-29T00:00:00</orderdate>
-        <total>45.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>OCEAN</id>
-    <name>Océano Atlántico Ltda.</name>
-    <address>Ing.Gustavo Moncada 8585, Piso 20-A</address>
-    <city>Buenos Aires</city>
-    <postalcode>1010</postalcode>
-    <country>Argentina</country>
-    <phone>(1) 135-5333</phone>
-    <fax>(1) 135-5535</fax>
-    <orders>
-      <order>
-        <id>10409</id>
-        <orderdate>1997-01-09T00:00:00</orderdate>
-        <total>319.20</total>
-      </order>
-      <order>
-        <id>10531</id>
-        <orderdate>1997-05-08T00:00:00</orderdate>
-        <total>110.00</total>
-      </order>
-      <order>
-        <id>10898</id>
-        <orderdate>1998-02-20T00:00:00</orderdate>
-        <total>30.00</total>
-      </order>
-      <order>
-        <id>10958</id>
-        <orderdate>1998-03-18T00:00:00</orderdate>
-        <total>781.00</total>
-      </order>
-      <order>
-        <id>10986</id>
-        <orderdate>1998-03-30T00:00:00</orderdate>
-        <total>2220.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>OLDWO</id>
-    <name>Old World Delicatessen</name>
-    <address>2743 Bering St.</address>
-    <city>Anchorage</city>
-    <region>AK</region>
-    <postalcode>99508</postalcode>
-    <country>USA</country>
-    <phone>(907) 555-7584</phone>
-    <fax>(907) 555-2880</fax>
-    <orders>
-      <order>
-        <id>10260</id>
-        <orderdate>1996-07-19T00:00:00</orderdate>
-        <total>1504.65</total>
-      </order>
-      <order>
-        <id>10305</id>
-        <orderdate>1996-09-13T00:00:00</orderdate>
-        <total>3741.30</total>
-      </order>
-      <order>
-        <id>10338</id>
-        <orderdate>1996-10-25T00:00:00</orderdate>
-        <total>934.50</total>
-      </order>
-      <order>
-        <id>10441</id>
-        <orderdate>1997-02-10T00:00:00</orderdate>
-        <total>1755.00</total>
-      </order>
-      <order>
-        <id>10594</id>
-        <orderdate>1997-07-09T00:00:00</orderdate>
-        <total>565.50</total>
-      </order>
-      <order>
-        <id>10680</id>
-        <orderdate>1997-09-24T00:00:00</orderdate>
-        <total>1261.88</total>
-      </order>
-      <order>
-        <id>10706</id>
-        <orderdate>1997-10-16T00:00:00</orderdate>
-        <total>1893.00</total>
-      </order>
-      <order>
-        <id>10855</id>
-        <orderdate>1998-01-27T00:00:00</orderdate>
-        <total>2227.89</total>
-      </order>
-      <order>
-        <id>10965</id>
-        <orderdate>1998-03-20T00:00:00</orderdate>
-        <total>848.00</total>
-      </order>
-      <order>
-        <id>11034</id>
-        <orderdate>1998-04-20T00:00:00</orderdate>
-        <total>539.40</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>OTTIK</id>
-    <name>Ottilies Käseladen</name>
-    <address>Mehrheimerstr. 369</address>
-    <city>Köln</city>
-    <postalcode>50739</postalcode>
-    <country>Germany</country>
-    <phone>0221-0644327</phone>
-    <fax>0221-0765721</fax>
-    <orders>
-      <order>
-        <id>10407</id>
-        <orderdate>1997-01-07T00:00:00</orderdate>
-        <total>1194.00</total>
-      </order>
-      <order>
-        <id>10508</id>
-        <orderdate>1997-04-16T00:00:00</orderdate>
-        <total>240.00</total>
-      </order>
-      <order>
-        <id>10554</id>
-        <orderdate>1997-05-30T00:00:00</orderdate>
-        <total>1728.52</total>
-      </order>
-      <order>
-        <id>10580</id>
-        <orderdate>1997-06-26T00:00:00</orderdate>
-        <total>1013.74</total>
-      </order>
-      <order>
-        <id>10684</id>
-        <orderdate>1997-09-26T00:00:00</orderdate>
-        <total>1768.00</total>
-      </order>
-      <order>
-        <id>10766</id>
-        <orderdate>1997-12-05T00:00:00</orderdate>
-        <total>2310.00</total>
-      </order>
-      <order>
-        <id>10833</id>
-        <orderdate>1998-01-15T00:00:00</orderdate>
-        <total>906.93</total>
-      </order>
-      <order>
-        <id>10999</id>
-        <orderdate>1998-04-03T00:00:00</orderdate>
-        <total>1197.95</total>
-      </order>
-      <order>
-        <id>11020</id>
-        <orderdate>1998-04-14T00:00:00</orderdate>
-        <total>632.40</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>PARIS</id>
-    <name>Paris spécialités</name>
-    <address>265, boulevard Charonne</address>
-    <city>Paris</city>
-    <postalcode>75012</postalcode>
-    <country>France</country>
-    <phone>(1) 42.34.22.66</phone>
-    <fax>(1) 42.34.22.77</fax>
-    <orders />
-  </customer>
-  <customer>
-    <id>PERIC</id>
-    <name>Pericles Comidas clásicas</name>
-    <address>Calle Dr.Jorge Cash 321</address>
-    <city>México D.F.</city>
-    <postalcode>05033</postalcode>
-    <country>Mexico</country>
-    <phone>(5) 552-3745</phone>
-    <fax>(5) 545-3745</fax>
-    <orders>
-      <order>
-        <id>10322</id>
-        <orderdate>1996-10-04T00:00:00</orderdate>
-        <total>112.00</total>
-      </order>
-      <order>
-        <id>10354</id>
-        <orderdate>1996-11-14T00:00:00</orderdate>
-        <total>568.80</total>
-      </order>
-      <order>
-        <id>10474</id>
-        <orderdate>1997-03-13T00:00:00</orderdate>
-        <total>1249.10</total>
-      </order>
-      <order>
-        <id>10502</id>
-        <orderdate>1997-04-10T00:00:00</orderdate>
-        <total>816.30</total>
-      </order>
-      <order>
-        <id>10995</id>
-        <orderdate>1998-04-02T00:00:00</orderdate>
-        <total>1196.00</total>
-      </order>
-      <order>
-        <id>11073</id>
-        <orderdate>1998-05-05T00:00:00</orderdate>
-        <total>300.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>PICCO</id>
-    <name>Piccolo und mehr</name>
-    <address>Geislweg 14</address>
-    <city>Salzburg</city>
-    <postalcode>5020</postalcode>
-    <country>Austria</country>
-    <phone>6562-9722</phone>
-    <fax>6562-9723</fax>
-    <orders>
-      <order>
-        <id>10353</id>
-        <orderdate>1996-11-13T00:00:00</orderdate>
-        <total>8593.28</total>
-      </order>
-      <order>
-        <id>10392</id>
-        <orderdate>1996-12-24T00:00:00</orderdate>
-        <total>1440.00</total>
-      </order>
-      <order>
-        <id>10427</id>
-        <orderdate>1997-01-27T00:00:00</orderdate>
-        <total>651.00</total>
-      </order>
-      <order>
-        <id>10489</id>
-        <orderdate>1997-03-28T00:00:00</orderdate>
-        <total>439.20</total>
-      </order>
-      <order>
-        <id>10530</id>
-        <orderdate>1997-05-08T00:00:00</orderdate>
-        <total>4180.00</total>
-      </order>
-      <order>
-        <id>10597</id>
-        <orderdate>1997-07-11T00:00:00</orderdate>
-        <total>718.08</total>
-      </order>
-      <order>
-        <id>10686</id>
-        <orderdate>1997-09-30T00:00:00</orderdate>
-        <total>1404.45</total>
-      </order>
-      <order>
-        <id>10747</id>
-        <orderdate>1997-11-19T00:00:00</orderdate>
-        <total>1912.85</total>
-      </order>
-      <order>
-        <id>10844</id>
-        <orderdate>1998-01-21T00:00:00</orderdate>
-        <total>735.00</total>
-      </order>
-      <order>
-        <id>11053</id>
-        <orderdate>1998-04-27T00:00:00</orderdate>
-        <total>3055.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>PRINI</id>
-    <name>Princesa Isabel Vinhos</name>
-    <address>Estrada da saúde n. 58</address>
-    <city>Lisboa</city>
-    <postalcode>1756</postalcode>
-    <country>Portugal</country>
-    <phone>(1) 356-5634</phone>
-    <orders>
-      <order>
-        <id>10336</id>
-        <orderdate>1996-10-23T00:00:00</orderdate>
-        <total>285.12</total>
-      </order>
-      <order>
-        <id>10397</id>
-        <orderdate>1996-12-27T00:00:00</orderdate>
-        <total>716.72</total>
-      </order>
-      <order>
-        <id>10433</id>
-        <orderdate>1997-02-03T00:00:00</orderdate>
-        <total>851.20</total>
-      </order>
-      <order>
-        <id>10477</id>
-        <orderdate>1997-03-17T00:00:00</orderdate>
-        <total>558.00</total>
-      </order>
-      <order>
-        <id>10808</id>
-        <orderdate>1998-01-01T00:00:00</orderdate>
-        <total>1411.00</total>
-      </order>
-      <order>
-        <id>11007</id>
-        <orderdate>1998-04-08T00:00:00</orderdate>
-        <total>2633.90</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>QUEDE</id>
-    <name>Que Delícia</name>
-    <address>Rua da Panificadora, 12</address>
-    <city>Rio de Janeiro</city>
-    <region>RJ</region>
-    <postalcode>02389-673</postalcode>
-    <country>Brazil</country>
-    <phone>(21) 555-4252</phone>
-    <fax>(21) 555-4545</fax>
-    <orders>
-      <order>
-        <id>10261</id>
-        <orderdate>1996-07-19T00:00:00</orderdate>
-        <total>448.00</total>
-      </order>
-      <order>
-        <id>10291</id>
-        <orderdate>1996-08-27T00:00:00</orderdate>
-        <total>497.52</total>
-      </order>
-      <order>
-        <id>10379</id>
-        <orderdate>1996-12-11T00:00:00</orderdate>
-        <total>863.28</total>
-      </order>
-      <order>
-        <id>10421</id>
-        <orderdate>1997-01-21T00:00:00</orderdate>
-        <total>1194.27</total>
-      </order>
-      <order>
-        <id>10587</id>
-        <orderdate>1997-07-02T00:00:00</orderdate>
-        <total>807.38</total>
-      </order>
-      <order>
-        <id>10647</id>
-        <orderdate>1997-08-27T00:00:00</orderdate>
-        <total>636.00</total>
-      </order>
-      <order>
-        <id>10720</id>
-        <orderdate>1997-10-28T00:00:00</orderdate>
-        <total>550.00</total>
-      </order>
-      <order>
-        <id>10794</id>
-        <orderdate>1997-12-24T00:00:00</orderdate>
-        <total>314.76</total>
-      </order>
-      <order>
-        <id>10989</id>
-        <orderdate>1998-03-31T00:00:00</orderdate>
-        <total>1353.60</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>QUEEN</id>
-    <name>Queen Cozinha</name>
-    <address>Alameda dos Canàrios, 891</address>
-    <city>São Paulo</city>
-    <region>SP</region>
-    <postalcode>05487-020</postalcode>
-    <country>Brazil</country>
-    <phone>(11) 555-1189</phone>
-    <orders>
-      <order>
-        <id>10372</id>
-        <orderdate>1996-12-04T00:00:00</orderdate>
-        <total>9210.90</total>
-      </order>
-      <order>
-        <id>10406</id>
-        <orderdate>1997-01-07T00:00:00</orderdate>
-        <total>1830.78</total>
-      </order>
-      <order>
-        <id>10487</id>
-        <orderdate>1997-03-26T00:00:00</orderdate>
-        <total>889.70</total>
-      </order>
-      <order>
-        <id>10637</id>
-        <orderdate>1997-08-19T00:00:00</orderdate>
-        <total>2761.94</total>
-      </order>
-      <order>
-        <id>10659</id>
-        <orderdate>1997-09-05T00:00:00</orderdate>
-        <total>1227.02</total>
-      </order>
-      <order>
-        <id>10704</id>
-        <orderdate>1997-10-14T00:00:00</orderdate>
-        <total>595.50</total>
-      </order>
-      <order>
-        <id>10728</id>
-        <orderdate>1997-11-04T00:00:00</orderdate>
-        <total>1296.75</total>
-      </order>
-      <order>
-        <id>10786</id>
-        <orderdate>1997-12-19T00:00:00</orderdate>
-        <total>1531.08</total>
-      </order>
-      <order>
-        <id>10868</id>
-        <orderdate>1998-02-04T00:00:00</orderdate>
-        <total>1920.60</total>
-      </order>
-      <order>
-        <id>10913</id>
-        <orderdate>1998-02-26T00:00:00</orderdate>
-        <total>768.75</total>
-      </order>
-      <order>
-        <id>10914</id>
-        <orderdate>1998-02-27T00:00:00</orderdate>
-        <total>537.50</total>
-      </order>
-      <order>
-        <id>10961</id>
-        <orderdate>1998-03-19T00:00:00</orderdate>
-        <total>1119.90</total>
-      </order>
-      <order>
-        <id>11068</id>
-        <orderdate>1998-05-04T00:00:00</orderdate>
-        <total>2027.08</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>QUICK</id>
-    <name>QUICK-Stop</name>
-    <address>Taucherstraße 10</address>
-    <city>Cunewalde</city>
-    <postalcode>01307</postalcode>
-    <country>Germany</country>
-    <phone>0372-035188</phone>
-    <orders>
-      <order>
-        <id>10273</id>
-        <orderdate>1996-08-05T00:00:00</orderdate>
-        <total>2037.28</total>
-      </order>
-      <order>
-        <id>10285</id>
-        <orderdate>1996-08-20T00:00:00</orderdate>
-        <total>1743.36</total>
-      </order>
-      <order>
-        <id>10286</id>
-        <orderdate>1996-08-21T00:00:00</orderdate>
-        <total>3016.00</total>
-      </order>
-      <order>
-        <id>10313</id>
-        <orderdate>1996-09-24T00:00:00</orderdate>
-        <total>182.40</total>
-      </order>
-      <order>
-        <id>10345</id>
-        <orderdate>1996-11-04T00:00:00</orderdate>
-        <total>2924.80</total>
-      </order>
-      <order>
-        <id>10361</id>
-        <orderdate>1996-11-22T00:00:00</orderdate>
-        <total>2046.24</total>
-      </order>
-      <order>
-        <id>10418</id>
-        <orderdate>1997-01-17T00:00:00</orderdate>
-        <total>1814.80</total>
-      </order>
-      <order>
-        <id>10451</id>
-        <orderdate>1997-02-19T00:00:00</orderdate>
-        <total>3849.66</total>
-      </order>
-      <order>
-        <id>10515</id>
-        <orderdate>1997-04-23T00:00:00</orderdate>
-        <total>9921.30</total>
-      </order>
-      <order>
-        <id>10527</id>
-        <orderdate>1997-05-05T00:00:00</orderdate>
-        <total>1503.00</total>
-      </order>
-      <order>
-        <id>10540</id>
-        <orderdate>1997-05-19T00:00:00</orderdate>
-        <total>10191.70</total>
-      </order>
-      <order>
-        <id>10549</id>
-        <orderdate>1997-05-27T00:00:00</orderdate>
-        <total>3554.28</total>
-      </order>
-      <order>
-        <id>10588</id>
-        <orderdate>1997-07-03T00:00:00</orderdate>
-        <total>3120.00</total>
-      </order>
-      <order>
-        <id>10658</id>
-        <orderdate>1997-09-05T00:00:00</orderdate>
-        <total>4464.60</total>
-      </order>
-      <order>
-        <id>10691</id>
-        <orderdate>1997-10-03T00:00:00</orderdate>
-        <total>10164.80</total>
-      </order>
-      <order>
-        <id>10694</id>
-        <orderdate>1997-10-06T00:00:00</orderdate>
-        <total>4825.00</total>
-      </order>
-      <order>
-        <id>10721</id>
-        <orderdate>1997-10-29T00:00:00</orderdate>
-        <total>923.88</total>
-      </order>
-      <order>
-        <id>10745</id>
-        <orderdate>1997-11-18T00:00:00</orderdate>
-        <total>4529.80</total>
-      </order>
-      <order>
-        <id>10765</id>
-        <orderdate>1997-12-04T00:00:00</orderdate>
-        <total>1515.60</total>
-      </order>
-      <order>
-        <id>10788</id>
-        <orderdate>1997-12-22T00:00:00</orderdate>
-        <total>731.50</total>
-      </order>
-      <order>
-        <id>10845</id>
-        <orderdate>1998-01-21T00:00:00</orderdate>
-        <total>3812.70</total>
-      </order>
-      <order>
-        <id>10865</id>
-        <orderdate>1998-02-02T00:00:00</orderdate>
-        <total>16387.50</total>
-      </order>
-      <order>
-        <id>10878</id>
-        <orderdate>1998-02-10T00:00:00</orderdate>
-        <total>1539.00</total>
-      </order>
-      <order>
-        <id>10938</id>
-        <orderdate>1998-03-10T00:00:00</orderdate>
-        <total>2731.88</total>
-      </order>
-      <order>
-        <id>10962</id>
-        <orderdate>1998-03-19T00:00:00</orderdate>
-        <total>3584.00</total>
-      </order>
-      <order>
-        <id>10991</id>
-        <orderdate>1998-04-01T00:00:00</orderdate>
-        <total>2296.00</total>
-      </order>
-      <order>
-        <id>10996</id>
-        <orderdate>1998-04-02T00:00:00</orderdate>
-        <total>560.00</total>
-      </order>
-      <order>
-        <id>11021</id>
-        <orderdate>1998-04-14T00:00:00</orderdate>
-        <total>6306.24</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>RANCH</id>
-    <name>Rancho grande</name>
-    <address>Av.del Libertador 900</address>
-    <city>Buenos Aires</city>
-    <postalcode>1010</postalcode>
-    <country>Argentina</country>
-    <phone>(1) 123-5555</phone>
-    <fax>(1) 123-5556</fax>
-    <orders>
-      <order>
-        <id>10448</id>
-        <orderdate>1997-02-17T00:00:00</orderdate>
-        <total>443.40</total>
-      </order>
-      <order>
-        <id>10716</id>
-        <orderdate>1997-10-24T00:00:00</orderdate>
-        <total>706.00</total>
-      </order>
-      <order>
-        <id>10828</id>
-        <orderdate>1998-01-13T00:00:00</orderdate>
-        <total>932.00</total>
-      </order>
-      <order>
-        <id>10916</id>
-        <orderdate>1998-02-27T00:00:00</orderdate>
-        <total>686.70</total>
-      </order>
-      <order>
-        <id>11019</id>
-        <orderdate>1998-04-13T00:00:00</orderdate>
-        <total>76.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>RATTC</id>
-    <name>Rattlesnake Canyon Grocery</name>
-    <address>2817 Milton Dr.</address>
-    <city>Albuquerque</city>
-    <region>NM</region>
-    <postalcode>87110</postalcode>
-    <country>USA</country>
-    <phone>(505) 555-5939</phone>
-    <fax>(505) 555-3620</fax>
-    <orders>
-      <order>
-        <id>10262</id>
-        <orderdate>1996-07-22T00:00:00</orderdate>
-        <total>584.00</total>
-      </order>
-      <order>
-        <id>10272</id>
-        <orderdate>1996-08-02T00:00:00</orderdate>
-        <total>1456.00</total>
-      </order>
-      <order>
-        <id>10294</id>
-        <orderdate>1996-08-30T00:00:00</orderdate>
-        <total>1887.60</total>
-      </order>
-      <order>
-        <id>10314</id>
-        <orderdate>1996-09-25T00:00:00</orderdate>
-        <total>2094.30</total>
-      </order>
-      <order>
-        <id>10316</id>
-        <orderdate>1996-09-27T00:00:00</orderdate>
-        <total>2835.00</total>
-      </order>
-      <order>
-        <id>10346</id>
-        <orderdate>1996-11-05T00:00:00</orderdate>
-        <total>1618.88</total>
-      </order>
-      <order>
-        <id>10401</id>
-        <orderdate>1997-01-01T00:00:00</orderdate>
-        <total>3868.60</total>
-      </order>
-      <order>
-        <id>10479</id>
-        <orderdate>1997-03-19T00:00:00</orderdate>
-        <total>10495.60</total>
-      </order>
-      <order>
-        <id>10564</id>
-        <orderdate>1997-06-10T00:00:00</orderdate>
-        <total>1234.05</total>
-      </order>
-      <order>
-        <id>10569</id>
-        <orderdate>1997-06-16T00:00:00</orderdate>
-        <total>890.00</total>
-      </order>
-      <order>
-        <id>10598</id>
-        <orderdate>1997-07-14T00:00:00</orderdate>
-        <total>2388.50</total>
-      </order>
-      <order>
-        <id>10761</id>
-        <orderdate>1997-12-02T00:00:00</orderdate>
-        <total>507.00</total>
-      </order>
-      <order>
-        <id>10820</id>
-        <orderdate>1998-01-07T00:00:00</orderdate>
-        <total>1140.00</total>
-      </order>
-      <order>
-        <id>10852</id>
-        <orderdate>1998-01-26T00:00:00</orderdate>
-        <total>2984.00</total>
-      </order>
-      <order>
-        <id>10889</id>
-        <orderdate>1998-02-16T00:00:00</orderdate>
-        <total>11380.00</total>
-      </order>
-      <order>
-        <id>10988</id>
-        <orderdate>1998-03-31T00:00:00</orderdate>
-        <total>3574.80</total>
-      </order>
-      <order>
-        <id>11000</id>
-        <orderdate>1998-04-06T00:00:00</orderdate>
-        <total>903.75</total>
-      </order>
-      <order>
-        <id>11077</id>
-        <orderdate>1998-05-06T00:00:00</orderdate>
-        <total>1255.72</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>REGGC</id>
-    <name>Reggiani Caseifici</name>
-    <address>Strada Provinciale 124</address>
-    <city>Reggio Emilia</city>
-    <postalcode>42100</postalcode>
-    <country>Italy</country>
-    <phone>0522-556721</phone>
-    <fax>0522-556722</fax>
-    <orders>
-      <order>
-        <id>10288</id>
-        <orderdate>1996-08-23T00:00:00</orderdate>
-        <total>80.10</total>
-      </order>
-      <order>
-        <id>10428</id>
-        <orderdate>1997-01-28T00:00:00</orderdate>
-        <total>192.00</total>
-      </order>
-      <order>
-        <id>10443</id>
-        <orderdate>1997-02-12T00:00:00</orderdate>
-        <total>517.44</total>
-      </order>
-      <order>
-        <id>10562</id>
-        <orderdate>1997-06-09T00:00:00</orderdate>
-        <total>488.70</total>
-      </order>
-      <order>
-        <id>10586</id>
-        <orderdate>1997-07-02T00:00:00</orderdate>
-        <total>23.80</total>
-      </order>
-      <order>
-        <id>10655</id>
-        <orderdate>1997-09-03T00:00:00</orderdate>
-        <total>154.40</total>
-      </order>
-      <order>
-        <id>10727</id>
-        <orderdate>1997-11-03T00:00:00</orderdate>
-        <total>1624.50</total>
-      </order>
-      <order>
-        <id>10812</id>
-        <orderdate>1998-01-02T00:00:00</orderdate>
-        <total>1692.80</total>
-      </order>
-      <order>
-        <id>10908</id>
-        <orderdate>1998-02-26T00:00:00</orderdate>
-        <total>663.10</total>
-      </order>
-      <order>
-        <id>10942</id>
-        <orderdate>1998-03-11T00:00:00</orderdate>
-        <total>560.00</total>
-      </order>
-      <order>
-        <id>11010</id>
-        <orderdate>1998-04-09T00:00:00</orderdate>
-        <total>645.00</total>
-      </order>
-      <order>
-        <id>11062</id>
-        <orderdate>1998-04-30T00:00:00</orderdate>
-        <total>406.40</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>RICAR</id>
-    <name>Ricardo Adocicados</name>
-    <address>Av.Copacabana, 267</address>
-    <city>Rio de Janeiro</city>
-    <region>RJ</region>
-    <postalcode>02389-890</postalcode>
-    <country>Brazil</country>
-    <phone>(21) 555-3412</phone>
-    <orders>
-      <order>
-        <id>10287</id>
-        <orderdate>1996-08-22T00:00:00</orderdate>
-        <total>819.00</total>
-      </order>
-      <order>
-        <id>10299</id>
-        <orderdate>1996-09-06T00:00:00</orderdate>
-        <total>349.50</total>
-      </order>
-      <order>
-        <id>10447</id>
-        <orderdate>1997-02-14T00:00:00</orderdate>
-        <total>914.40</total>
-      </order>
-      <order>
-        <id>10481</id>
-        <orderdate>1997-03-20T00:00:00</orderdate>
-        <total>1472.00</total>
-      </order>
-      <order>
-        <id>10563</id>
-        <orderdate>1997-06-10T00:00:00</orderdate>
-        <total>965.00</total>
-      </order>
-      <order>
-        <id>10622</id>
-        <orderdate>1997-08-06T00:00:00</orderdate>
-        <total>560.00</total>
-      </order>
-      <order>
-        <id>10648</id>
-        <orderdate>1997-08-28T00:00:00</orderdate>
-        <total>372.38</total>
-      </order>
-      <order>
-        <id>10813</id>
-        <orderdate>1998-01-05T00:00:00</orderdate>
-        <total>602.40</total>
-      </order>
-      <order>
-        <id>10851</id>
-        <orderdate>1998-01-26T00:00:00</orderdate>
-        <total>2603.00</total>
-      </order>
-      <order>
-        <id>10877</id>
-        <orderdate>1998-02-09T00:00:00</orderdate>
-        <total>1955.12</total>
-      </order>
-      <order>
-        <id>11059</id>
-        <orderdate>1998-04-29T00:00:00</orderdate>
-        <total>1838.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>RICSU</id>
-    <name>Richter Supermarkt</name>
-    <address>Grenzacherweg 237</address>
-    <city>Genève</city>
-    <postalcode>1203</postalcode>
-    <country>Switzerland</country>
-    <phone>0897-034214</phone>
-    <orders>
-      <order>
-        <id>10255</id>
-        <orderdate>1996-07-12T00:00:00</orderdate>
-        <total>2490.50</total>
-      </order>
-      <order>
-        <id>10419</id>
-        <orderdate>1997-01-20T00:00:00</orderdate>
-        <total>2097.60</total>
-      </order>
-      <order>
-        <id>10537</id>
-        <orderdate>1997-05-14T00:00:00</orderdate>
-        <total>1823.80</total>
-      </order>
-      <order>
-        <id>10666</id>
-        <orderdate>1997-09-12T00:00:00</orderdate>
-        <total>4666.94</total>
-      </order>
-      <order>
-        <id>10751</id>
-        <orderdate>1997-11-24T00:00:00</orderdate>
-        <total>1631.48</total>
-      </order>
-      <order>
-        <id>10758</id>
-        <orderdate>1997-11-28T00:00:00</orderdate>
-        <total>1644.60</total>
-      </order>
-      <order>
-        <id>10931</id>
-        <orderdate>1998-03-06T00:00:00</orderdate>
-        <total>799.20</total>
-      </order>
-      <order>
-        <id>10951</id>
-        <orderdate>1998-03-16T00:00:00</orderdate>
-        <total>458.76</total>
-      </order>
-      <order>
-        <id>11033</id>
-        <orderdate>1998-04-17T00:00:00</orderdate>
-        <total>3232.80</total>
-      </order>
-      <order>
-        <id>11075</id>
-        <orderdate>1998-05-06T00:00:00</orderdate>
-        <total>498.10</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>ROMEY</id>
-    <name>Romero y tomillo</name>
-    <address>Gran Vía, 1</address>
-    <city>Madrid</city>
-    <postalcode>28001</postalcode>
-    <country>Spain</country>
-    <phone>(91) 745 6200</phone>
-    <fax>(91) 745 6210</fax>
-    <orders>
-      <order>
-        <id>10281</id>
-        <orderdate>1996-08-14T00:00:00</orderdate>
-        <total>86.50</total>
-      </order>
-      <order>
-        <id>10282</id>
-        <orderdate>1996-08-15T00:00:00</orderdate>
-        <total>155.40</total>
-      </order>
-      <order>
-        <id>10306</id>
-        <orderdate>1996-09-16T00:00:00</orderdate>
-        <total>498.50</total>
-      </order>
-      <order>
-        <id>10917</id>
-        <orderdate>1998-03-02T00:00:00</orderdate>
-        <total>365.89</total>
-      </order>
-      <order>
-        <id>11013</id>
-        <orderdate>1998-04-09T00:00:00</orderdate>
-        <total>361.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>SANTG</id>
-    <name>Santé Gourmet</name>
-    <address>Erling Skakkes gate 78</address>
-    <city>Stavern</city>
-    <postalcode>4110</postalcode>
-    <country>Norway</country>
-    <phone>07-98 92 35</phone>
-    <fax>07-98 92 47</fax>
-    <orders>
-      <order>
-        <id>10387</id>
-        <orderdate>1996-12-18T00:00:00</orderdate>
-        <total>1058.40</total>
-      </order>
-      <order>
-        <id>10520</id>
-        <orderdate>1997-04-29T00:00:00</orderdate>
-        <total>200.00</total>
-      </order>
-      <order>
-        <id>10639</id>
-        <orderdate>1997-08-20T00:00:00</orderdate>
-        <total>500.00</total>
-      </order>
-      <order>
-        <id>10831</id>
-        <orderdate>1998-01-14T00:00:00</orderdate>
-        <total>2684.40</total>
-      </order>
-      <order>
-        <id>10909</id>
-        <orderdate>1998-02-26T00:00:00</orderdate>
-        <total>670.00</total>
-      </order>
-      <order>
-        <id>11015</id>
-        <orderdate>1998-04-10T00:00:00</orderdate>
-        <total>622.35</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>SAVEA</id>
-    <name>Save-a-lot Markets</name>
-    <address>187 Suffolk Ln.</address>
-    <city>Boise</city>
-    <region>ID</region>
-    <postalcode>83720</postalcode>
-    <country>USA</country>
-    <phone>(208) 555-8097</phone>
-    <orders>
-      <order>
-        <id>10324</id>
-        <orderdate>1996-10-08T00:00:00</orderdate>
-        <total>5275.72</total>
-      </order>
-      <order>
-        <id>10393</id>
-        <orderdate>1996-12-25T00:00:00</orderdate>
-        <total>2556.95</total>
-      </order>
-      <order>
-        <id>10398</id>
-        <orderdate>1996-12-30T00:00:00</orderdate>
-        <total>2505.60</total>
-      </order>
-      <order>
-        <id>10440</id>
-        <orderdate>1997-02-10T00:00:00</orderdate>
-        <total>4924.14</total>
-      </order>
-      <order>
-        <id>10452</id>
-        <orderdate>1997-02-20T00:00:00</orderdate>
-        <total>2018.50</total>
-      </order>
-      <order>
-        <id>10510</id>
-        <orderdate>1997-04-18T00:00:00</orderdate>
-        <total>4707.54</total>
-      </order>
-      <order>
-        <id>10555</id>
-        <orderdate>1997-06-02T00:00:00</orderdate>
-        <total>2944.40</total>
-      </order>
-      <order>
-        <id>10603</id>
-        <orderdate>1997-07-18T00:00:00</orderdate>
-        <total>1483.00</total>
-      </order>
-      <order>
-        <id>10607</id>
-        <orderdate>1997-07-22T00:00:00</orderdate>
-        <total>6475.40</total>
-      </order>
-      <order>
-        <id>10612</id>
-        <orderdate>1997-07-28T00:00:00</orderdate>
-        <total>6375.00</total>
-      </order>
-      <order>
-        <id>10627</id>
-        <orderdate>1997-08-11T00:00:00</orderdate>
-        <total>1185.75</total>
-      </order>
-      <order>
-        <id>10657</id>
-        <orderdate>1997-09-04T00:00:00</orderdate>
-        <total>4371.60</total>
-      </order>
-      <order>
-        <id>10678</id>
-        <orderdate>1997-09-23T00:00:00</orderdate>
-        <total>5256.50</total>
-      </order>
-      <order>
-        <id>10700</id>
-        <orderdate>1997-10-10T00:00:00</orderdate>
-        <total>1638.40</total>
-      </order>
-      <order>
-        <id>10711</id>
-        <orderdate>1997-10-21T00:00:00</orderdate>
-        <total>4451.70</total>
-      </order>
-      <order>
-        <id>10713</id>
-        <orderdate>1997-10-22T00:00:00</orderdate>
-        <total>2827.90</total>
-      </order>
-      <order>
-        <id>10714</id>
-        <orderdate>1997-10-22T00:00:00</orderdate>
-        <total>2205.75</total>
-      </order>
-      <order>
-        <id>10722</id>
-        <orderdate>1997-10-29T00:00:00</orderdate>
-        <total>1570.00</total>
-      </order>
-      <order>
-        <id>10748</id>
-        <orderdate>1997-11-20T00:00:00</orderdate>
-        <total>2196.00</total>
-      </order>
-      <order>
-        <id>10757</id>
-        <orderdate>1997-11-27T00:00:00</orderdate>
-        <total>3082.00</total>
-      </order>
-      <order>
-        <id>10815</id>
-        <orderdate>1998-01-05T00:00:00</orderdate>
-        <total>40.00</total>
-      </order>
-      <order>
-        <id>10847</id>
-        <orderdate>1998-01-22T00:00:00</orderdate>
-        <total>4931.92</total>
-      </order>
-      <order>
-        <id>10882</id>
-        <orderdate>1998-02-11T00:00:00</orderdate>
-        <total>892.64</total>
-      </order>
-      <order>
-        <id>10894</id>
-        <orderdate>1998-02-18T00:00:00</orderdate>
-        <total>2753.10</total>
-      </order>
-      <order>
-        <id>10941</id>
-        <orderdate>1998-03-11T00:00:00</orderdate>
-        <total>4011.75</total>
-      </order>
-      <order>
-        <id>10983</id>
-        <orderdate>1998-03-27T00:00:00</orderdate>
-        <total>720.90</total>
-      </order>
-      <order>
-        <id>10984</id>
-        <orderdate>1998-03-30T00:00:00</orderdate>
-        <total>1809.75</total>
-      </order>
-      <order>
-        <id>11002</id>
-        <orderdate>1998-04-06T00:00:00</orderdate>
-        <total>1811.10</total>
-      </order>
-      <order>
-        <id>11030</id>
-        <orderdate>1998-04-17T00:00:00</orderdate>
-        <total>12615.05</total>
-      </order>
-      <order>
-        <id>11031</id>
-        <orderdate>1998-04-17T00:00:00</orderdate>
-        <total>2393.50</total>
-      </order>
-      <order>
-        <id>11064</id>
-        <orderdate>1998-05-01T00:00:00</orderdate>
-        <total>4330.40</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>SEVES</id>
-    <name>Seven Seas Imports</name>
-    <address>90 Wadhurst Rd.</address>
-    <city>London</city>
-    <postalcode>OX15 4NB</postalcode>
-    <country>UK</country>
-    <phone>(171) 555-1717</phone>
-    <fax>(171) 555-5646</fax>
-    <orders>
-      <order>
-        <id>10359</id>
-        <orderdate>1996-11-21T00:00:00</orderdate>
-        <total>3471.68</total>
-      </order>
-      <order>
-        <id>10377</id>
-        <orderdate>1996-12-09T00:00:00</orderdate>
-        <total>863.60</total>
-      </order>
-      <order>
-        <id>10388</id>
-        <orderdate>1996-12-19T00:00:00</orderdate>
-        <total>1228.80</total>
-      </order>
-      <order>
-        <id>10472</id>
-        <orderdate>1997-03-12T00:00:00</orderdate>
-        <total>1036.80</total>
-      </order>
-      <order>
-        <id>10523</id>
-        <orderdate>1997-05-01T00:00:00</orderdate>
-        <total>2444.31</total>
-      </order>
-      <order>
-        <id>10547</id>
-        <orderdate>1997-05-23T00:00:00</orderdate>
-        <total>1792.80</total>
-      </order>
-      <order>
-        <id>10800</id>
-        <orderdate>1997-12-26T00:00:00</orderdate>
-        <total>1468.94</total>
-      </order>
-      <order>
-        <id>10804</id>
-        <orderdate>1997-12-30T00:00:00</orderdate>
-        <total>2278.40</total>
-      </order>
-      <order>
-        <id>10869</id>
-        <orderdate>1998-02-04T00:00:00</orderdate>
-        <total>1630.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>SIMOB</id>
-    <name>Simons bistro</name>
-    <address>Vinbæltet 34</address>
-    <city>København</city>
-    <postalcode>1734</postalcode>
-    <country>Denmark</country>
-    <phone>31 12 34 56</phone>
-    <fax>31 13 35 57</fax>
-    <orders>
-      <order>
-        <id>10341</id>
-        <orderdate>1996-10-29T00:00:00</orderdate>
-        <total>352.60</total>
-      </order>
-      <order>
-        <id>10417</id>
-        <orderdate>1997-01-16T00:00:00</orderdate>
-        <total>11188.40</total>
-      </order>
-      <order>
-        <id>10556</id>
-        <orderdate>1997-06-03T00:00:00</orderdate>
-        <total>835.20</total>
-      </order>
-      <order>
-        <id>10642</id>
-        <orderdate>1997-08-22T00:00:00</orderdate>
-        <total>696.00</total>
-      </order>
-      <order>
-        <id>10669</id>
-        <orderdate>1997-09-15T00:00:00</orderdate>
-        <total>570.00</total>
-      </order>
-      <order>
-        <id>10802</id>
-        <orderdate>1997-12-29T00:00:00</orderdate>
-        <total>2942.81</total>
-      </order>
-      <order>
-        <id>11074</id>
-        <orderdate>1998-05-06T00:00:00</orderdate>
-        <total>232.08</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>SPECD</id>
-    <name>Spécialités du monde</name>
-    <address>25, rue Lauriston</address>
-    <city>Paris</city>
-    <postalcode>75016</postalcode>
-    <country>France</country>
-    <phone>(1) 47.55.60.10</phone>
-    <fax>(1) 47.55.60.20</fax>
-    <orders>
-      <order>
-        <id>10738</id>
-        <orderdate>1997-11-12T00:00:00</orderdate>
-        <total>52.35</total>
-      </order>
-      <order>
-        <id>10907</id>
-        <orderdate>1998-02-25T00:00:00</orderdate>
-        <total>108.50</total>
-      </order>
-      <order>
-        <id>10964</id>
-        <orderdate>1998-03-20T00:00:00</orderdate>
-        <total>2052.50</total>
-      </order>
-      <order>
-        <id>11043</id>
-        <orderdate>1998-04-22T00:00:00</orderdate>
-        <total>210.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>SPLIR</id>
-    <name>Split Rail Beer &amp; Ale</name>
-    <address>P.O.Box 555</address>
-    <city>Lander</city>
-    <region>WY</region>
-    <postalcode>82520</postalcode>
-    <country>USA</country>
-    <phone>(307) 555-4680</phone>
-    <fax>(307) 555-6525</fax>
-    <orders>
-      <order>
-        <id>10271</id>
-        <orderdate>1996-08-01T00:00:00</orderdate>
-        <total>48.00</total>
-      </order>
-      <order>
-        <id>10329</id>
-        <orderdate>1996-10-15T00:00:00</orderdate>
-        <total>4578.43</total>
-      </order>
-      <order>
-        <id>10349</id>
-        <orderdate>1996-11-08T00:00:00</orderdate>
-        <total>141.60</total>
-      </order>
-      <order>
-        <id>10369</id>
-        <orderdate>1996-12-02T00:00:00</orderdate>
-        <total>2390.40</total>
-      </order>
-      <order>
-        <id>10385</id>
-        <orderdate>1996-12-17T00:00:00</orderdate>
-        <total>691.20</total>
-      </order>
-      <order>
-        <id>10432</id>
-        <orderdate>1997-01-31T00:00:00</orderdate>
-        <total>485.00</total>
-      </order>
-      <order>
-        <id>10756</id>
-        <orderdate>1997-11-27T00:00:00</orderdate>
-        <total>1990.00</total>
-      </order>
-      <order>
-        <id>10821</id>
-        <orderdate>1998-01-08T00:00:00</orderdate>
-        <total>678.00</total>
-      </order>
-      <order>
-        <id>10974</id>
-        <orderdate>1998-03-25T00:00:00</orderdate>
-        <total>439.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>SUPRD</id>
-    <name>Suprêmes délices</name>
-    <address>Boulevard Tirou, 255</address>
-    <city>Charleroi</city>
-    <postalcode>B-6000</postalcode>
-    <country>Belgium</country>
-    <phone>(071) 23 67 22 20</phone>
-    <fax>(071) 23 67 22 21</fax>
-    <orders>
-      <order>
-        <id>10252</id>
-        <orderdate>1996-07-09T00:00:00</orderdate>
-        <total>3597.90</total>
-      </order>
-      <order>
-        <id>10302</id>
-        <orderdate>1996-09-10T00:00:00</orderdate>
-        <total>2708.80</total>
-      </order>
-      <order>
-        <id>10458</id>
-        <orderdate>1997-02-26T00:00:00</orderdate>
-        <total>3891.00</total>
-      </order>
-      <order>
-        <id>10463</id>
-        <orderdate>1997-03-04T00:00:00</orderdate>
-        <total>713.30</total>
-      </order>
-      <order>
-        <id>10475</id>
-        <orderdate>1997-03-14T00:00:00</orderdate>
-        <total>1505.18</total>
-      </order>
-      <order>
-        <id>10767</id>
-        <orderdate>1997-12-05T00:00:00</orderdate>
-        <total>28.00</total>
-      </order>
-      <order>
-        <id>10841</id>
-        <orderdate>1998-01-20T00:00:00</orderdate>
-        <total>4581.00</total>
-      </order>
-      <order>
-        <id>10846</id>
-        <orderdate>1998-01-22T00:00:00</orderdate>
-        <total>1112.00</total>
-      </order>
-      <order>
-        <id>10885</id>
-        <orderdate>1998-02-12T00:00:00</orderdate>
-        <total>1209.00</total>
-      </order>
-      <order>
-        <id>10930</id>
-        <orderdate>1998-03-06T00:00:00</orderdate>
-        <total>2255.50</total>
-      </order>
-      <order>
-        <id>11035</id>
-        <orderdate>1998-04-20T00:00:00</orderdate>
-        <total>1754.50</total>
-      </order>
-      <order>
-        <id>11038</id>
-        <orderdate>1998-04-21T00:00:00</orderdate>
-        <total>732.60</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>THEBI</id>
-    <name>The Big Cheese</name>
-    <address>89 Jefferson Way, Suite 2</address>
-    <city>Portland</city>
-    <region>OR</region>
-    <postalcode>97201</postalcode>
-    <country>USA</country>
-    <phone>(503) 555-3612</phone>
-    <orders>
-      <order>
-        <id>10310</id>
-        <orderdate>1996-09-20T00:00:00</orderdate>
-        <total>336.00</total>
-      </order>
-      <order>
-        <id>10708</id>
-        <orderdate>1997-10-17T00:00:00</orderdate>
-        <total>180.40</total>
-      </order>
-      <order>
-        <id>10805</id>
-        <orderdate>1997-12-30T00:00:00</orderdate>
-        <total>2775.00</total>
-      </order>
-      <order>
-        <id>10992</id>
-        <orderdate>1998-04-01T00:00:00</orderdate>
-        <total>69.60</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>THECR</id>
-    <name>The Cracker Box</name>
-    <address>55 Grizzly Peak Rd.</address>
-    <city>Butte</city>
-    <region>MT</region>
-    <postalcode>59801</postalcode>
-    <country>USA</country>
-    <phone>(406) 555-5834</phone>
-    <fax>(406) 555-8083</fax>
-    <orders>
-      <order>
-        <id>10624</id>
-        <orderdate>1997-08-07T00:00:00</orderdate>
-        <total>1393.24</total>
-      </order>
-      <order>
-        <id>10775</id>
-        <orderdate>1997-12-12T00:00:00</orderdate>
-        <total>228.00</total>
-      </order>
-      <order>
-        <id>11003</id>
-        <orderdate>1998-04-06T00:00:00</orderdate>
-        <total>326.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>TOMSP</id>
-    <name>Toms Spezialitäten</name>
-    <address>Luisenstr. 48</address>
-    <city>Münster</city>
-    <postalcode>44087</postalcode>
-    <country>Germany</country>
-    <phone>0251-031259</phone>
-    <fax>0251-035695</fax>
-    <orders>
-      <order>
-        <id>10438</id>
-        <orderdate>1997-02-06T00:00:00</orderdate>
-        <total>454.00</total>
-      </order>
-      <order>
-        <id>10446</id>
-        <orderdate>1997-02-14T00:00:00</orderdate>
-        <total>246.24</total>
-      </order>
-      <order>
-        <id>10548</id>
-        <orderdate>1997-05-26T00:00:00</orderdate>
-        <total>240.10</total>
-      </order>
-      <order>
-        <id>10608</id>
-        <orderdate>1997-07-23T00:00:00</orderdate>
-        <total>1064.00</total>
-      </order>
-      <order>
-        <id>10967</id>
-        <orderdate>1998-03-23T00:00:00</orderdate>
-        <total>910.40</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>TORTU</id>
-    <name>Tortuga Restaurante</name>
-    <address>Avda.Azteca 123</address>
-    <city>México D.F.</city>
-    <postalcode>05033</postalcode>
-    <country>Mexico</country>
-    <phone>(5) 555-2933</phone>
-    <orders>
-      <order>
-        <id>10276</id>
-        <orderdate>1996-08-08T00:00:00</orderdate>
-        <total>420.00</total>
-      </order>
-      <order>
-        <id>10293</id>
-        <orderdate>1996-08-29T00:00:00</orderdate>
-        <total>848.70</total>
-      </order>
-      <order>
-        <id>10304</id>
-        <orderdate>1996-09-12T00:00:00</orderdate>
-        <total>954.40</total>
-      </order>
-      <order>
-        <id>10319</id>
-        <orderdate>1996-10-02T00:00:00</orderdate>
-        <total>1191.20</total>
-      </order>
-      <order>
-        <id>10518</id>
-        <orderdate>1997-04-25T00:00:00</orderdate>
-        <total>4150.05</total>
-      </order>
-      <order>
-        <id>10576</id>
-        <orderdate>1997-06-23T00:00:00</orderdate>
-        <total>838.45</total>
-      </order>
-      <order>
-        <id>10676</id>
-        <orderdate>1997-09-22T00:00:00</orderdate>
-        <total>534.85</total>
-      </order>
-      <order>
-        <id>10842</id>
-        <orderdate>1998-01-20T00:00:00</orderdate>
-        <total>975.00</total>
-      </order>
-      <order>
-        <id>10915</id>
-        <orderdate>1998-02-27T00:00:00</orderdate>
-        <total>539.50</total>
-      </order>
-      <order>
-        <id>11069</id>
-        <orderdate>1998-05-04T00:00:00</orderdate>
-        <total>360.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>TRADH</id>
-    <name>Tradição Hipermercados</name>
-    <address>Av.Inês de Castro, 414</address>
-    <city>São Paulo</city>
-    <region>SP</region>
-    <postalcode>05634-030</postalcode>
-    <country>Brazil</country>
-    <phone>(11) 555-2167</phone>
-    <fax>(11) 555-2168</fax>
-    <orders>
-      <order>
-        <id>10249</id>
-        <orderdate>1996-07-05T00:00:00</orderdate>
-        <total>1863.40</total>
-      </order>
-      <order>
-        <id>10292</id>
-        <orderdate>1996-08-28T00:00:00</orderdate>
-        <total>1296.00</total>
-      </order>
-      <order>
-        <id>10496</id>
-        <orderdate>1997-04-04T00:00:00</orderdate>
-        <total>190.00</total>
-      </order>
-      <order>
-        <id>10606</id>
-        <orderdate>1997-07-22T00:00:00</orderdate>
-        <total>1130.40</total>
-      </order>
-      <order>
-        <id>10830</id>
-        <orderdate>1998-01-13T00:00:00</orderdate>
-        <total>1974.00</total>
-      </order>
-      <order>
-        <id>10834</id>
-        <orderdate>1998-01-15T00:00:00</orderdate>
-        <total>1432.71</total>
-      </order>
-      <order>
-        <id>10839</id>
-        <orderdate>1998-01-19T00:00:00</orderdate>
-        <total>827.55</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>TRAIH</id>
-    <name>Trail's Head Gourmet Provisioners</name>
-    <address>722 DaVinci Blvd.</address>
-    <city>Kirkland</city>
-    <region>WA</region>
-    <postalcode>98034</postalcode>
-    <country>USA</country>
-    <phone>(206) 555-8257</phone>
-    <fax>(206) 555-2174</fax>
-    <orders>
-      <order>
-        <id>10574</id>
-        <orderdate>1997-06-19T00:00:00</orderdate>
-        <total>764.30</total>
-      </order>
-      <order>
-        <id>10577</id>
-        <orderdate>1997-06-23T00:00:00</orderdate>
-        <total>569.00</total>
-      </order>
-      <order>
-        <id>10822</id>
-        <orderdate>1998-01-08T00:00:00</orderdate>
-        <total>237.90</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>VAFFE</id>
-    <name>Vaffeljernet</name>
-    <address>Smagsløget 45</address>
-    <city>Århus</city>
-    <postalcode>8200</postalcode>
-    <country>Denmark</country>
-    <phone>86 21 32 43</phone>
-    <fax>86 22 33 44</fax>
-    <orders>
-      <order>
-        <id>10367</id>
-        <orderdate>1996-11-28T00:00:00</orderdate>
-        <total>834.20</total>
-      </order>
-      <order>
-        <id>10399</id>
-        <orderdate>1996-12-31T00:00:00</orderdate>
-        <total>1765.60</total>
-      </order>
-      <order>
-        <id>10465</id>
-        <orderdate>1997-03-05T00:00:00</orderdate>
-        <total>2518.00</total>
-      </order>
-      <order>
-        <id>10591</id>
-        <orderdate>1997-07-07T00:00:00</orderdate>
-        <total>812.50</total>
-      </order>
-      <order>
-        <id>10602</id>
-        <orderdate>1997-07-17T00:00:00</orderdate>
-        <total>48.75</total>
-      </order>
-      <order>
-        <id>10688</id>
-        <orderdate>1997-10-01T00:00:00</orderdate>
-        <total>3160.60</total>
-      </order>
-      <order>
-        <id>10744</id>
-        <orderdate>1997-11-17T00:00:00</orderdate>
-        <total>736.00</total>
-      </order>
-      <order>
-        <id>10769</id>
-        <orderdate>1997-12-08T00:00:00</orderdate>
-        <total>1684.28</total>
-      </order>
-      <order>
-        <id>10921</id>
-        <orderdate>1998-03-03T00:00:00</orderdate>
-        <total>1936.00</total>
-      </order>
-      <order>
-        <id>10946</id>
-        <orderdate>1998-03-12T00:00:00</orderdate>
-        <total>1407.50</total>
-      </order>
-      <order>
-        <id>10994</id>
-        <orderdate>1998-04-02T00:00:00</orderdate>
-        <total>940.50</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>VICTE</id>
-    <name>Victuailles en stock</name>
-    <address>2, rue du Commerce</address>
-    <city>Lyon</city>
-    <postalcode>69004</postalcode>
-    <country>France</country>
-    <phone>78.32.54.86</phone>
-    <fax>78.32.54.87</fax>
-    <orders>
-      <order>
-        <id>10251</id>
-        <orderdate>1996-07-08T00:00:00</orderdate>
-        <total>654.06</total>
-      </order>
-      <order>
-        <id>10334</id>
-        <orderdate>1996-10-21T00:00:00</orderdate>
-        <total>144.80</total>
-      </order>
-      <order>
-        <id>10450</id>
-        <orderdate>1997-02-19T00:00:00</orderdate>
-        <total>425.12</total>
-      </order>
-      <order>
-        <id>10459</id>
-        <orderdate>1997-02-27T00:00:00</orderdate>
-        <total>1659.20</total>
-      </order>
-      <order>
-        <id>10478</id>
-        <orderdate>1997-03-18T00:00:00</orderdate>
-        <total>471.20</total>
-      </order>
-      <order>
-        <id>10546</id>
-        <orderdate>1997-05-23T00:00:00</orderdate>
-        <total>2812.00</total>
-      </order>
-      <order>
-        <id>10806</id>
-        <orderdate>1997-12-31T00:00:00</orderdate>
-        <total>439.60</total>
-      </order>
-      <order>
-        <id>10814</id>
-        <orderdate>1998-01-05T00:00:00</orderdate>
-        <total>1788.45</total>
-      </order>
-      <order>
-        <id>10843</id>
-        <orderdate>1998-01-21T00:00:00</orderdate>
-        <total>159.00</total>
-      </order>
-      <order>
-        <id>10850</id>
-        <orderdate>1998-01-23T00:00:00</orderdate>
-        <total>629.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>VINET</id>
-    <name>Vins et alcools Chevalier</name>
-    <address>59 rue de l'Abbaye</address>
-    <city>Reims</city>
-    <postalcode>51100</postalcode>
-    <country>France</country>
-    <phone>26.47.15.10</phone>
-    <fax>26.47.15.11</fax>
-    <orders>
-      <order>
-        <id>10274</id>
-        <orderdate>1996-08-06T00:00:00</orderdate>
-        <total>538.60</total>
-      </order>
-      <order>
-        <id>10295</id>
-        <orderdate>1996-09-02T00:00:00</orderdate>
-        <total>121.60</total>
-      </order>
-      <order>
-        <id>10737</id>
-        <orderdate>1997-11-11T00:00:00</orderdate>
-        <total>139.80</total>
-      </order>
-      <order>
-        <id>10739</id>
-        <orderdate>1997-11-12T00:00:00</orderdate>
-        <total>240.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>WANDK</id>
-    <name>Die Wandernde Kuh</name>
-    <address>Adenauerallee 900</address>
-    <city>Stuttgart</city>
-    <postalcode>70563</postalcode>
-    <country>Germany</country>
-    <phone>0711-020361</phone>
-    <fax>0711-035428</fax>
-    <orders>
-      <order>
-        <id>10301</id>
-        <orderdate>1996-09-09T00:00:00</orderdate>
-        <total>755.00</total>
-      </order>
-      <order>
-        <id>10312</id>
-        <orderdate>1996-09-23T00:00:00</orderdate>
-        <total>1614.80</total>
-      </order>
-      <order>
-        <id>10348</id>
-        <orderdate>1996-11-07T00:00:00</orderdate>
-        <total>363.60</total>
-      </order>
-      <order>
-        <id>10356</id>
-        <orderdate>1996-11-18T00:00:00</orderdate>
-        <total>1106.40</total>
-      </order>
-      <order>
-        <id>10513</id>
-        <orderdate>1997-04-22T00:00:00</orderdate>
-        <total>1942.00</total>
-      </order>
-      <order>
-        <id>10632</id>
-        <orderdate>1997-08-14T00:00:00</orderdate>
-        <total>589.00</total>
-      </order>
-      <order>
-        <id>10640</id>
-        <orderdate>1997-08-21T00:00:00</orderdate>
-        <total>708.75</total>
-      </order>
-      <order>
-        <id>10651</id>
-        <orderdate>1997-09-01T00:00:00</orderdate>
-        <total>397.80</total>
-      </order>
-      <order>
-        <id>10668</id>
-        <orderdate>1997-09-15T00:00:00</orderdate>
-        <total>625.28</total>
-      </order>
-      <order>
-        <id>11046</id>
-        <orderdate>1998-04-23T00:00:00</orderdate>
-        <total>1485.80</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>WARTH</id>
-    <name>Wartian Herkku</name>
-    <address>Torikatu 38</address>
-    <city>Oulu</city>
-    <postalcode>90110</postalcode>
-    <country>Finland</country>
-    <phone>981-443655</phone>
-    <fax>981-443655</fax>
-    <orders>
-      <order>
-        <id>10266</id>
-        <orderdate>1996-07-26T00:00:00</orderdate>
-        <total>346.56</total>
-      </order>
-      <order>
-        <id>10270</id>
-        <orderdate>1996-08-01T00:00:00</orderdate>
-        <total>1376.00</total>
-      </order>
-      <order>
-        <id>10320</id>
-        <orderdate>1996-10-03T00:00:00</orderdate>
-        <total>516.00</total>
-      </order>
-      <order>
-        <id>10333</id>
-        <orderdate>1996-10-18T00:00:00</orderdate>
-        <total>877.20</total>
-      </order>
-      <order>
-        <id>10412</id>
-        <orderdate>1997-01-13T00:00:00</orderdate>
-        <total>334.80</total>
-      </order>
-      <order>
-        <id>10416</id>
-        <orderdate>1997-01-16T00:00:00</orderdate>
-        <total>720.00</total>
-      </order>
-      <order>
-        <id>10437</id>
-        <orderdate>1997-02-05T00:00:00</orderdate>
-        <total>393.00</total>
-      </order>
-      <order>
-        <id>10455</id>
-        <orderdate>1997-02-24T00:00:00</orderdate>
-        <total>2684.00</total>
-      </order>
-      <order>
-        <id>10526</id>
-        <orderdate>1997-05-05T00:00:00</orderdate>
-        <total>1151.40</total>
-      </order>
-      <order>
-        <id>10553</id>
-        <orderdate>1997-05-30T00:00:00</orderdate>
-        <total>1546.30</total>
-      </order>
-      <order>
-        <id>10583</id>
-        <orderdate>1997-06-30T00:00:00</orderdate>
-        <total>2237.50</total>
-      </order>
-      <order>
-        <id>10636</id>
-        <orderdate>1997-08-19T00:00:00</orderdate>
-        <total>629.50</total>
-      </order>
-      <order>
-        <id>10750</id>
-        <orderdate>1997-11-21T00:00:00</orderdate>
-        <total>1590.56</total>
-      </order>
-      <order>
-        <id>10781</id>
-        <orderdate>1997-12-17T00:00:00</orderdate>
-        <total>975.88</total>
-      </order>
-      <order>
-        <id>11025</id>
-        <orderdate>1998-04-15T00:00:00</orderdate>
-        <total>270.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>WELLI</id>
-    <name>Wellington Importadora</name>
-    <address>Rua do Mercado, 12</address>
-    <city>Resende</city>
-    <region>SP</region>
-    <postalcode>08737-363</postalcode>
-    <country>Brazil</country>
-    <phone>(14) 555-8122</phone>
-    <orders>
-      <order>
-        <id>10256</id>
-        <orderdate>1996-07-15T00:00:00</orderdate>
-        <total>517.80</total>
-      </order>
-      <order>
-        <id>10420</id>
-        <orderdate>1997-01-21T00:00:00</orderdate>
-        <total>1707.84</total>
-      </order>
-      <order>
-        <id>10585</id>
-        <orderdate>1997-07-01T00:00:00</orderdate>
-        <total>142.50</total>
-      </order>
-      <order>
-        <id>10644</id>
-        <orderdate>1997-08-25T00:00:00</orderdate>
-        <total>1371.80</total>
-      </order>
-      <order>
-        <id>10803</id>
-        <orderdate>1997-12-30T00:00:00</orderdate>
-        <total>1193.01</total>
-      </order>
-      <order>
-        <id>10809</id>
-        <orderdate>1998-01-01T00:00:00</orderdate>
-        <total>140.00</total>
-      </order>
-      <order>
-        <id>10900</id>
-        <orderdate>1998-02-20T00:00:00</orderdate>
-        <total>33.75</total>
-      </order>
-      <order>
-        <id>10905</id>
-        <orderdate>1998-02-24T00:00:00</orderdate>
-        <total>342.00</total>
-      </order>
-      <order>
-        <id>10935</id>
-        <orderdate>1998-03-09T00:00:00</orderdate>
-        <total>619.50</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>WHITC</id>
-    <name>White Clover Markets</name>
-    <address>305 - 14th Ave.S.Suite 3B</address>
-    <city>Seattle</city>
-    <region>WA</region>
-    <postalcode>98128</postalcode>
-    <country>USA</country>
-    <phone>(206) 555-4112</phone>
-    <fax>(206) 555-4115</fax>
-    <orders>
-      <order>
-        <id>10269</id>
-        <orderdate>1996-07-31T00:00:00</orderdate>
-        <total>642.20</total>
-      </order>
-      <order>
-        <id>10344</id>
-        <orderdate>1996-11-01T00:00:00</orderdate>
-        <total>2296.00</total>
-      </order>
-      <order>
-        <id>10469</id>
-        <orderdate>1997-03-10T00:00:00</orderdate>
-        <total>956.68</total>
-      </order>
-      <order>
-        <id>10483</id>
-        <orderdate>1997-03-24T00:00:00</orderdate>
-        <total>668.80</total>
-      </order>
-      <order>
-        <id>10504</id>
-        <orderdate>1997-04-11T00:00:00</orderdate>
-        <total>1388.50</total>
-      </order>
-      <order>
-        <id>10596</id>
-        <orderdate>1997-07-11T00:00:00</orderdate>
-        <total>1180.88</total>
-      </order>
-      <order>
-        <id>10693</id>
-        <orderdate>1997-10-06T00:00:00</orderdate>
-        <total>2071.20</total>
-      </order>
-      <order>
-        <id>10696</id>
-        <orderdate>1997-10-08T00:00:00</orderdate>
-        <total>996.00</total>
-      </order>
-      <order>
-        <id>10723</id>
-        <orderdate>1997-10-30T00:00:00</orderdate>
-        <total>468.45</total>
-      </order>
-      <order>
-        <id>10740</id>
-        <orderdate>1997-11-13T00:00:00</orderdate>
-        <total>1416.00</total>
-      </order>
-      <order>
-        <id>10861</id>
-        <orderdate>1998-01-30T00:00:00</orderdate>
-        <total>3523.40</total>
-      </order>
-      <order>
-        <id>10904</id>
-        <orderdate>1998-02-24T00:00:00</orderdate>
-        <total>1924.25</total>
-      </order>
-      <order>
-        <id>11032</id>
-        <orderdate>1998-04-17T00:00:00</orderdate>
-        <total>8902.50</total>
-      </order>
-      <order>
-        <id>11066</id>
-        <orderdate>1998-05-01T00:00:00</orderdate>
-        <total>928.75</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>WILMK</id>
-    <name>Wilman Kala</name>
-    <address>Keskuskatu 45</address>
-    <city>Helsinki</city>
-    <postalcode>21240</postalcode>
-    <country>Finland</country>
-    <phone>90-224 8858</phone>
-    <fax>90-224 8858</fax>
-    <orders>
-      <order>
-        <id>10248</id>
-        <orderdate>1996-07-04T00:00:00</orderdate>
-        <total>440.00</total>
-      </order>
-      <order>
-        <id>10615</id>
-        <orderdate>1997-07-30T00:00:00</orderdate>
-        <total>120.00</total>
-      </order>
-      <order>
-        <id>10673</id>
-        <orderdate>1997-09-18T00:00:00</orderdate>
-        <total>412.35</total>
-      </order>
-      <order>
-        <id>10695</id>
-        <orderdate>1997-10-07T00:00:00</orderdate>
-        <total>642.00</total>
-      </order>
-      <order>
-        <id>10873</id>
-        <orderdate>1998-02-06T00:00:00</orderdate>
-        <total>336.80</total>
-      </order>
-      <order>
-        <id>10879</id>
-        <orderdate>1998-02-10T00:00:00</orderdate>
-        <total>611.30</total>
-      </order>
-      <order>
-        <id>10910</id>
-        <orderdate>1998-02-26T00:00:00</orderdate>
-        <total>452.90</total>
-      </order>
-      <order>
-        <id>11005</id>
-        <orderdate>1998-04-07T00:00:00</orderdate>
-        <total>586.00</total>
-      </order>
-    </orders>
-  </customer>
-  <customer>
-    <id>WOLZA</id>
-    <name>Wolski Zajazd</name>
-    <address>ul.Filtrowa 68</address>
-    <city>Warszawa</city>
-    <postalcode>01-012</postalcode>
-    <country>Poland</country>
-    <phone>(26) 642-7012</phone>
-    <fax>(26) 642-7012</fax>
-    <orders>
-      <order>
-        <id>10374</id>
-        <orderdate>1996-12-05T00:00:00</orderdate>
-        <total>459.00</total>
-      </order>
-      <order>
-        <id>10611</id>
-        <orderdate>1997-07-25T00:00:00</orderdate>
-        <total>808.00</total>
-      </order>
-      <order>
-        <id>10792</id>
-        <orderdate>1997-12-23T00:00:00</orderdate>
-        <total>399.85</total>
-      </order>
-      <order>
-        <id>10870</id>
-        <orderdate>1998-02-04T00:00:00</orderdate>
-        <total>160.00</total>
-      </order>
-      <order>
-        <id>10906</id>
-        <orderdate>1998-02-25T00:00:00</orderdate>
-        <total>427.50</total>
-      </order>
-      <order>
-        <id>10998</id>
-        <orderdate>1998-04-03T00:00:00</orderdate>
-        <total>686.00</total>
-      </order>
-      <order>
-        <id>11044</id>
-        <orderdate>1998-04-23T00:00:00</orderdate>
-        <total>591.60</total>
-      </order>
-    </orders>
-  </customer>
-</customers>";
-        #endregion
+        public List<Product> GetProductList() => Products.ProductList;
 
-        private List<Product> productList =
-            new List<Product> {
-                new Product { ProductID = 1, ProductName = "Chai", Category = "Beverages", UnitPrice = 18.0000M, UnitsInStock = 39 },
-                new Product { ProductID = 2, ProductName = "Chang", Category = "Beverages", UnitPrice = 19.0000M, UnitsInStock = 17 },
-                new Product { ProductID = 3, ProductName = "Aniseed Syrup", Category = "Condiments", UnitPrice = 10.0000M, UnitsInStock = 13 },
-                new Product { ProductID = 4, ProductName = "Chef Anton's Cajun Seasoning", Category = "Condiments", UnitPrice = 22.0000M, UnitsInStock = 53 },
-                new Product { ProductID = 5, ProductName = "Chef Anton's Gumbo Mix", Category = "Condiments", UnitPrice = 21.3500M, UnitsInStock = 0 },
-                new Product { ProductID = 6, ProductName = "Grandma's Boysenberry Spread", Category = "Condiments", UnitPrice = 25.0000M, UnitsInStock = 120 },
-                new Product { ProductID = 7, ProductName = "Uncle Bob's Organic Dried Pears", Category = "Produce", UnitPrice = 30.0000M, UnitsInStock = 15 },
-                new Product { ProductID = 8, ProductName = "Northwoods Cranberry Sauce", Category = "Condiments", UnitPrice = 40.0000M, UnitsInStock = 6 },
-                new Product { ProductID = 9, ProductName = "Mishi Kobe Niku", Category = "Meat/Poultry", UnitPrice = 97.0000M, UnitsInStock = 29 },
-                new Product { ProductID = 10, ProductName = "Ikura", Category = "Seafood", UnitPrice = 31.0000M, UnitsInStock = 31 },
-                new Product { ProductID = 11, ProductName = "Queso Cabrales", Category = "Dairy Products", UnitPrice = 21.0000M, UnitsInStock = 22 },
-                new Product { ProductID = 12, ProductName = "Queso Manchego La Pastora", Category = "Dairy Products", UnitPrice = 38.0000M, UnitsInStock = 86 },
-                new Product { ProductID = 13, ProductName = "Konbu", Category = "Seafood", UnitPrice = 6.0000M, UnitsInStock = 24 },
-                new Product { ProductID = 14, ProductName = "Tofu", Category = "Produce", UnitPrice = 23.2500M, UnitsInStock = 35 },
-                new Product { ProductID = 15, ProductName = "Genen Shouyu", Category = "Condiments", UnitPrice = 15.5000M, UnitsInStock = 39 },
-                new Product { ProductID = 16, ProductName = "Pavlova", Category = "Confections", UnitPrice = 17.4500M, UnitsInStock = 29 },
-                new Product { ProductID = 17, ProductName = "Alice Mutton", Category = "Meat/Poultry", UnitPrice = 39.0000M, UnitsInStock = 0 },
-                new Product { ProductID = 18, ProductName = "Carnarvon Tigers", Category = "Seafood", UnitPrice = 62.5000M, UnitsInStock = 42 },
-                new Product { ProductID = 19, ProductName = "Teatime Chocolate Biscuits", Category = "Confections", UnitPrice = 9.2000M, UnitsInStock = 25 },
-                new Product { ProductID = 20, ProductName = "Sir Rodney's Marmalade", Category = "Confections", UnitPrice = 81.0000M, UnitsInStock = 40 },
-                new Product { ProductID = 21, ProductName = "Sir Rodney's Scones", Category = "Confections", UnitPrice = 10.0000M, UnitsInStock = 3 },
-                new Product { ProductID = 22, ProductName = "Gustaf's Knäckebröd", Category = "Grains/Cereals", UnitPrice = 21.0000M, UnitsInStock = 104 },
-                new Product { ProductID = 23, ProductName = "Tunnbröd", Category = "Grains/Cereals", UnitPrice = 9.0000M, UnitsInStock = 61 },
-                new Product { ProductID = 24, ProductName = "Guaraná Fantástica", Category = "Beverages", UnitPrice = 4.5000M, UnitsInStock = 20 },
-                new Product { ProductID = 25, ProductName = "NuNuCa Nuß-Nougat-Creme", Category = "Confections", UnitPrice = 14.0000M, UnitsInStock = 76 },
-                new Product { ProductID = 26, ProductName = "Gumbär Gummibärchen", Category = "Confections", UnitPrice = 31.2300M, UnitsInStock = 15 },
-                new Product { ProductID = 27, ProductName = "Schoggi Schokolade", Category = "Confections", UnitPrice = 43.9000M, UnitsInStock = 49 },
-                new Product { ProductID = 28, ProductName = "Rössle Sauerkraut", Category = "Produce", UnitPrice = 45.6000M, UnitsInStock = 26 },
-                new Product { ProductID = 29, ProductName = "Thüringer Rostbratwurst", Category = "Meat/Poultry", UnitPrice = 123.7900M, UnitsInStock = 0 },
-                new Product { ProductID = 30, ProductName = "Nord-Ost Matjeshering", Category = "Seafood", UnitPrice = 25.8900M, UnitsInStock = 10 },
-                new Product { ProductID = 31, ProductName = "Gorgonzola Telino", Category = "Dairy Products", UnitPrice = 12.5000M, UnitsInStock = 0 },
-                new Product { ProductID = 32, ProductName = "Mascarpone Fabioli", Category = "Dairy Products", UnitPrice = 32.0000M, UnitsInStock = 9 },
-                new Product { ProductID = 33, ProductName = "Geitost", Category = "Dairy Products", UnitPrice = 2.5000M, UnitsInStock = 112 },
-                new Product { ProductID = 34, ProductName = "Sasquatch Ale", Category = "Beverages", UnitPrice = 14.0000M, UnitsInStock = 111 },
-                new Product { ProductID = 35, ProductName = "Steeleye Stout", Category = "Beverages", UnitPrice = 18.0000M, UnitsInStock = 20 },
-                new Product { ProductID = 36, ProductName = "Inlagd Sill", Category = "Seafood", UnitPrice = 19.0000M, UnitsInStock = 112 },
-                new Product { ProductID = 37, ProductName = "Gravad lax", Category = "Seafood", UnitPrice = 26.0000M, UnitsInStock = 11 },
-                new Product { ProductID = 38, ProductName = "Côte de Blaye", Category = "Beverages", UnitPrice = 263.5000M, UnitsInStock = 17 },
-                new Product { ProductID = 39, ProductName = "Chartreuse verte", Category = "Beverages", UnitPrice = 18.0000M, UnitsInStock = 69 },
-                new Product { ProductID = 40, ProductName = "Boston Crab Meat", Category = "Seafood", UnitPrice = 18.4000M, UnitsInStock = 123 },
-                new Product { ProductID = 41, ProductName = "Jack's New England Clam Chowder", Category = "Seafood", UnitPrice = 9.6500M, UnitsInStock = 85 },
-                new Product { ProductID = 42, ProductName = "Singaporean Hokkien Fried Mee", Category = "Grains/Cereals", UnitPrice = 14.0000M, UnitsInStock = 26 },
-                new Product { ProductID = 43, ProductName = "Ipoh Coffee", Category = "Beverages", UnitPrice = 46.0000M, UnitsInStock = 17 },
-                new Product { ProductID = 44, ProductName = "Gula Malacca", Category = "Condiments", UnitPrice = 19.4500M, UnitsInStock = 27 },
-                new Product { ProductID = 45, ProductName = "Rogede sild", Category = "Seafood", UnitPrice = 9.5000M, UnitsInStock = 5 },
-                new Product { ProductID = 46, ProductName = "Spegesild", Category = "Seafood", UnitPrice = 12.0000M, UnitsInStock = 95 },
-                new Product { ProductID = 47, ProductName = "Zaanse koeken", Category = "Confections", UnitPrice = 9.5000M, UnitsInStock = 36 },
-                new Product { ProductID = 48, ProductName = "Chocolade", Category = "Confections", UnitPrice = 12.7500M, UnitsInStock = 15 },
-                new Product { ProductID = 49, ProductName = "Maxilaku", Category = "Confections", UnitPrice = 20.0000M, UnitsInStock = 10 },
-                new Product { ProductID = 50, ProductName = "Valkoinen suklaa", Category = "Confections", UnitPrice = 16.2500M, UnitsInStock = 65 },
-                new Product { ProductID = 51, ProductName = "Manjimup Dried Apples", Category = "Produce", UnitPrice = 53.0000M, UnitsInStock = 20 },
-                new Product { ProductID = 52, ProductName = "Filo Mix", Category = "Grains/Cereals", UnitPrice = 7.0000M, UnitsInStock = 38 },
-                new Product { ProductID = 53, ProductName = "Perth Pasties", Category = "Meat/Poultry", UnitPrice = 32.8000M, UnitsInStock = 0 },
-                new Product { ProductID = 54, ProductName = "Tourtière", Category = "Meat/Poultry", UnitPrice = 7.4500M, UnitsInStock = 21 },
-                new Product { ProductID = 55, ProductName = "Pâté chinois", Category = "Meat/Poultry", UnitPrice = 24.0000M, UnitsInStock = 115 },
-                new Product { ProductID = 56, ProductName = "Gnocchi di nonna Alice", Category = "Grains/Cereals", UnitPrice = 38.0000M, UnitsInStock = 21 },
-                new Product { ProductID = 57, ProductName = "Ravioli Angelo", Category = "Grains/Cereals", UnitPrice = 19.5000M, UnitsInStock = 36 },
-                new Product { ProductID = 58, ProductName = "Escargots de Bourgogne", Category = "Seafood", UnitPrice = 13.2500M, UnitsInStock = 62 },
-                new Product { ProductID = 59, ProductName = "Raclette Courdavault", Category = "Dairy Products", UnitPrice = 55.0000M, UnitsInStock = 79 },
-                new Product { ProductID = 60, ProductName = "Camembert Pierrot", Category = "Dairy Products", UnitPrice = 34.0000M, UnitsInStock = 19 },
-                new Product { ProductID = 61, ProductName = "Sirop d'érable", Category = "Condiments", UnitPrice = 28.5000M, UnitsInStock = 113 },
-                new Product { ProductID = 62, ProductName = "Tarte au sucre", Category = "Confections", UnitPrice = 49.3000M, UnitsInStock = 17 },
-                new Product { ProductID = 63, ProductName = "Vegie-spread", Category = "Condiments", UnitPrice = 43.9000M, UnitsInStock = 24 },
-                new Product { ProductID = 64, ProductName = "Wimmers gute Semmelknödel", Category = "Grains/Cereals", UnitPrice = 33.2500M, UnitsInStock = 22 },
-                new Product { ProductID = 65, ProductName = "Louisiana Fiery Hot Pepper Sauce", Category = "Condiments", UnitPrice = 21.0500M, UnitsInStock = 76 },
-                new Product { ProductID = 66, ProductName = "Louisiana Hot Spiced Okra", Category = "Condiments", UnitPrice = 17.0000M, UnitsInStock = 4 },
-                new Product { ProductID = 67, ProductName = "Laughing Lumberjack Lager", Category = "Beverages", UnitPrice = 14.0000M, UnitsInStock = 52 },
-                new Product { ProductID = 68, ProductName = "Scottish Longbreads", Category = "Confections", UnitPrice = 12.5000M, UnitsInStock = 6 },
-                new Product { ProductID = 69, ProductName = "Gudbrandsdalsost", Category = "Dairy Products", UnitPrice = 36.0000M, UnitsInStock = 26 },
-                new Product { ProductID = 70, ProductName = "Outback Lager", Category = "Beverages", UnitPrice = 15.0000M, UnitsInStock = 15 },
-                new Product { ProductID = 71, ProductName = "Flotemysost", Category = "Dairy Products", UnitPrice = 21.5000M, UnitsInStock = 26 },
-                new Product { ProductID = 72, ProductName = "Mozzarella di Giovanni", Category = "Dairy Products", UnitPrice = 34.8000M, UnitsInStock = 14 },
-                new Product { ProductID = 73, ProductName = "Röd Kaviar", Category = "Seafood", UnitPrice = 15.0000M, UnitsInStock = 101 },
-                new Product { ProductID = 74, ProductName = "Longlife Tofu", Category = "Produce", UnitPrice = 10.0000M, UnitsInStock = 4 },
-                new Product { ProductID = 75, ProductName = "Rhönbräu Klosterbier", Category = "Beverages", UnitPrice = 7.7500M, UnitsInStock = 125 },
-                new Product { ProductID = 76, ProductName = "Lakkalikööri", Category = "Beverages", UnitPrice = 18.0000M, UnitsInStock = 57 },
-                new Product { ProductID = 77, ProductName = "Original Frankfurter grüne Soße", Category = "Condiments", UnitPrice = 13.0000M, UnitsInStock = 32 }
-            };
-
-        // TODO: Change to load from string stream:
-        private List<Customer> customerList = (from e in XDocument.Parse(customers_xml).
-                              Root.Elements("customer")
-                    select new Customer
-                    {
-                        CustomerID = (string)e.Element("id"),
-                        CompanyName = (string)e.Element("name"),
-                        Address = (string)e.Element("address"),
-                        City = (string)e.Element("city"),
-                        Region = (string)e.Element("region"),
-                        PostalCode = (string)e.Element("postalcode"),
-                        Country = (string)e.Element("country"),
-                        Phone = (string)e.Element("phone"),
-                        Fax = (string)e.Element("fax"),
-                        Orders = (
-                            from o in e.Elements("orders").Elements("order")
-                            select new Order
-                            {
-                                OrderID = (int)o.Element("id"),
-                                OrderDate = (DateTime)o.Element("orderdate"),
-                                Total = (decimal)o.Element("total")
-                            })
-                            .ToArray()
-                    })
-                    .ToList();
-
-
-        public List<Product> GetProductList() => productList;
-
-        public List<Customer> GetCustomerList() => customerList;
-
-        public Restrictions()
-        {
-        }
-
+        public List<Customer> GetCustomerList() => Customers.CustomerList;
 
         public int LowNumbers()
         {
